% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkRNAseq.R
\name{addMetaRNAseq}
\alias{addMetaRNAseq}
\alias{addMeta<-}
\alias{addMeta<-.RNAseq}
\title{Add metadata slot to an RNAseq object}
\usage{
addMetaRNAseq(value, name = deparse(substitute(value)), object = NULL)

addMeta(object) <- value

\method{addMeta}{RNAseq}(object) <- value
}
\arguments{
\item{value}{Vector of any type with length equal to the number of samples in the \code{object}}

\item{name}{String name for the metadata slot. Example: "age".
By default, the name of the object provided to value is used.

This will become the name of the slot AND the string that should be provided to \code{var}-type inputs of dittoSeq plotting funcitons.}

\item{object}{the \code{\linkS4class{RNAseq}} object. REQUIRED, unless \code{DEFAULT <- "object"} has been run.}
}
\value{
Outputs an \code{\linkS4class{RNAseq}} object with a new \code{@meta.data$'name'} slot.
}
\description{
Add metadata slot to an RNAseq object
}
\examples{

# Import mock data
myRNA <- RNAseq_mock

### Add a batch metadata
# Option 1, all in the funciton:
myRNA <- addMetaRNAseq(
    value = rep(1:2, each = 5),
    name = "batch",
    object = myRNA)
# Options 2 & 3, create variable beforehand, and provide that.
#     FOr both, the name of the variable will be name of the metadata slot.
batch <- rep(1:2, each = 5)
myRNA <- addMetaRNAseq(batch, object = myRNA)
# OR
addMeta(myRNA) <- batch

# Visualize batch on a PCA plot
#   Note: For RNAseq objects, reduction.use defaults to "pca"
#   so just dittoDimPlot("Sample", myRNA, size = 3) would work the same!
dittoDimPlot("batch", myRNA, reduction.use = "pca", size = 3)

}
\seealso{
\code{\link{importDESeq2}} for initial import of \code{RNAseq} data analyzed with DESeq2

\code{\link{addDimReduction}} and \code{\link{addPrcomp}} for adding dimensionality reductions

\code{\linkS4class{RNAseq}} for learning more about the \code{RNAseq} object type
}
\author{
Daniel Bunis
}
