% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{all_cells}
\alias{all_cells}
\title{Retrieves all the cells/samples}
\usage{
all_cells(object = DEFAULT)
}
\arguments{
\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}
}
\value{
Given a seurat or RNAseq object, will return the cell.names or samples slot.
}
\description{
Retrieves all the cells/samples
}
\examples{
pbmc <- Seurat::pbmc_small
all_cells("pbmc")
}
