% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoNeighborDiv.R
\name{calcNeighborMetadataDiversity}
\alias{calcNeighborMetadataDiversity}
\title{Shows data overlayed on a tsne, pca, or similar type of plot}
\usage{
calcNeighborMetadataDiversity(
  object,
  var,
  neighbors = .default_neighbors(object),
  distances,
  quantile = 0.9
)
}
\arguments{
\item{var}{String name of a "gene" or "metadata" (or "ident" for a Seurat \code{object}) to use for coloring the plots.
This is the data that will be displayed for each cell/sample. Discrete or continuous data both work.

Alternatively, a string vector naming multiple genes or metadata, OR a vector of the same length as there are cells/samples in the \code{object} which provides per-cell data directly.}

\item{neighbors}{a single string giving either the name of a Neighbors slot of the (Seurat) \code{object},
OR or matrix with cells in its rows and indexes of neighbors in its columns}

\item{distances}{not needed when \code{neighbors} is directed to a Neighbors slot of the (Seurat) \code{object},
Otherwise, must be given a matrix with cells in its rows and distance measures to each neighbor in its columns}
}
\value{
A named numeric vector of diversity counts, the same length as the number of cells in \code{object} which can be added to the object as cell metadata, and named by the cell names of the \code{object}.
}
\description{
Shows data overlayed on a tsne, pca, or similar type of plot
}
\details{
If given a Seurat \code{object} and \code{neighbors} is given (default) a string value representing a Neighbors object slot name.
It then extracts the \code{neighbors}-matrix and \code{distances}-matrix from this object.

Otherwise, it uses the \code{neighbors} and \code{distances} inputs for these purposes.

To calculate neighbors' Diversity:

First, the distance cutoff for neighbors deemed close-enough is determined based on the given \code{quantile} of \code{distance}-values.

Then, function then extracts the given \code{var} metadata from the object.

Finally, it loops through each cell (row) of the neighbors and distances matrices,
totaling the number of distinct var-values associated with the cell's neighbors that are within the threshold distance.  

Cell names are then added directly before the vector is output.
}
\examples{
example(importDittoBulk, echo = FALSE)
}
\seealso{
\code{\link{dittoDimPlot}} and \code{\link{dittoDimHex}} for additional details about the other options as these are the plotters used after diversity calculations complete.
}
\author{
Daniel Bunis
}
