% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{classof}
\alias{classof}
\title{Returns the class of an object when given the name of the object in "quotes"}
\usage{
classof(object = DEFAULT)
}
\arguments{
\item{object}{quoted "object" name}
}
\value{
Returns a the string name of the object's type.
}
\description{
Returns the class of an object when given the name of the object in "quotes"
}
\examples{
#' library(Seurat)
pbmc <- Seurat::pbmc_small
classof("pbmc")
}
