% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBBarPlot.R
\name{DBBarPlot}
\alias{DBBarPlot}
\title{Outputs a stacked bar plot to show the percent composition of samples or other cell groupings}
\usage{
DBBarPlot(var = "ident", object = DEFAULT, group.by = "Sample",
  cells.use = NULL, color.panel = dittoColors(),
  colors = c(1:length(color.panel)), do.hover = FALSE,
  theme = theme_classic(), xlab = NULL, ylab = "make",
  x.labels = NA, rotate.labels = TRUE, y.breaks = c(0, 0.5, 1),
  main = "make", sub = NULL, rename.groups = NA,
  legend.show = TRUE, legend.title = NULL,
  reorder.x = seq_along(meta.levels(group.by, object)))
}
\arguments{
\item{var}{Target Variable = values, OR a metadata in "quotes". REQUIRED. Length must be same  as the number of cells in the Seurat object or Samples in the RNAseq object}

\item{object}{the Seurat or RNAseq object to draw from = name of object in "quotes". REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{group.by}{"metadata" to use for separating values. REQUIRED (Default is to use a metadata named "Sample").}

\item{cells.use}{Cells to include: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. *USE* in object@cell.names[*USE*])}

\item{color.panel}{the set of colors to draw from}

\item{colors}{indexes / or order, of colors from color.panel to actual use}

\item{do.hover}{TRUE/FALSE. Default = FALSE.  If set to true, object will be converted to a ggplotly object so that data about individual bars will be displayed when you hover your cursor over them.  Data displayed will be the "counts" and "percentage of total".'data.hover' argument is not used with this plotting function.}

\item{theme}{Allows setting of a theme. Default = theme_classic() when nothing is provided.}

\item{xlab}{"character". The text title for the x axis.  NULL/blank by default.}

\item{ylab}{"character". The text title for the y axis.  Auto-generated by default.  Provide ylab = NULL to remove}

\item{x.labels}{Replacement x-axis labels to use instead of the identities of gorup.by}

\item{rotate.labels}{TRUE/FALSE. whether the labels should be rotated.  Default = FALSE = vertical labels.}

\item{y.breaks}{The numerical labels for the y axis.  Note: The percentages that build this figure will always add up to 1, so the plot will always go from 0 to 1.}

\item{main}{"character". Plot main title.}

\item{sub}{"character". Plot subtitle.}

\item{rename.groups}{new names for the identities of var.  Change to NULL to remove labeling altogether.}

\item{legend.show}{TRUE/FALSE. Whether the legend should be displayed. Default = TRUE.}

\item{legend.title}{Title for the legend.  Default = blank / NULL}

\item{reorder.x}{sequence of numbers from 1:length(meta.levels(group.by)) for providing a new order for the samples.  Default = alphabetical then numerical.}
}
\value{
Makes a plot where discrete data is shown on the y-axis as "percent of total" in a stacked barplot, grouped by sample, age, cluster, etc., on the x-axis.
}
\description{
Outputs a stacked bar plot to show the percent composition of samples or other cell groupings
}
\examples{
library(Seurat)
pbmc <- pbmc_small
DBBarPlot("ident", object = "pbmc", group.by = "orig.ident")
#For real data, you will have more distinct samples than this truncated dataset.
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
DBBarPlot("ident", group.by = "orig.ident")
}
