% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{DBDimPlot}
\alias{DBDimPlot}
\title{Show data overlayed on a tsne or pca or other reduction-type}
\usage{
DBDimPlot(var = "ident", object = DEFAULT, reduction.use = NA,
  dim.1 = 1, dim.2 = 2, theme = NA, size = 1, shape = 16,
  shapes = c(16, 15, 17, 23, 25, 8), legend.size = 5,
  legend.title = NULL, shape.legend.size = 5,
  shape.legend.title = NULL, main = NULL, sub = NULL,
  xlab = "make", ylab = "make", auto.title = T, cells.use = NULL,
  show.others = TRUE, ellipse = F, do.label = F, label.size = 5,
  highlight.labels = T, label.by = NULL, rename.groups = NA,
  low = "#F0E442", high = "#0072B2", range = NULL,
  color.panel = MYcolors, colors = 1:length(color.panel),
  do.letter = NA)
}
\arguments{
\item{var}{Target Variable = either values or a metadata (in "quotes"), gene (in "quotes"), or "ident"}

\item{object}{the Seurat or RNAseq object to work on}

\item{reduction.use}{"pca", "tsne", "ica", etc.  Default = tsne for Seurat objects, and pca for RNAseq objects}

\item{dim.1}{The component number to use on the x-axis.  Default = 1}

\item{dim.2}{The component number to use on the y-axis.  Default = 2}

\item{theme}{Allows setting of a theme. Default = theme_bw when nothing is provided.}

\item{size}{Number. Size of data points.  Default = 1.}

\item{shape}{Number for setting shape OR name of metadata to use for setting shape}

\item{shapes}{the shapes to use.  Default is a list of 6.  There are more, but not many of the default ggplot options are great.  I recommend using colors for any variable with 7+ options.}

\item{legend.size}{The size to increase the plotting of colors/letters legend shapes to (for discrete variable plotting)}

\item{legend.title}{For adding a title for the colors/letters legend.  It is set to NULL (off) by default.}

\item{shape.legend.size}{The size to increase the plotting of shapes legend shapes to.}

\item{shape.legend.title}{For adding a title for the shapes legend is a meta.data was given to 'shape' and multiple shapes were therefore used.  It is set to NULL (off) by default.}

\item{main}{plot title}

\item{sub}{plot subtitle}

\item{xlab}{label for y axes.  Default labels are generated if you do not give this a specific value.  To remove the labeling, set to NULL.}

\item{ylab}{label for y axes.  Default labels are generated if you do not give this a specific value.  To remove the labeling, set to NULL.}

\item{auto.title}{TRUE/FALSE = whether a default 'main' should be generated.}

\item{cells.use}{cells to show: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. *THIS*: object@cell.names[*THIS*])}

\item{show.others}{TRUE/FALSE. TRUE by default, whether other cells should be shown in the background}

\item{ellipse}{TRUE/FALSE. Whether the groups should be surrounded by an ellipse.}

\item{do.label}{TRUE/FALSE. Whether to add text labels at the center (median) of clusters for grouping vars}

\item{label.size}{Size of the the labels text}

\item{highlight.labels}{TRUE/FALSE. Whether the labels should have a box behind them}

\item{label.by}{To be removed.}

\item{rename.groups}{new names for the identities of var.  Change to NULL to remove labeling altogether.}

\item{low}{color for lowest values of var/range}

\item{high}{color for highest values of var/range}

\item{range}{limits for color scaling if var is a continuous value}

\item{color.panel}{a list of colors to be used for when plotting a discrete var.}

\item{colors}{indexes / order of colors from color.panel to use. USAGE= changing the order of how colors are linked to specific groups}

\item{do.letter}{TRUE/FALSE/NA. Whether letters should be added on top of the colored dots. For colorblindness compatibility.  NA by default, and if left that way, will be set to either TRUE or FALSE depending on the number of groups if a discrete var is given. For when there are lots of descrete variables, it can be hard to see by just color / shape.  NOTE: Lettering is incompatible with changing dots to shapes, so this will override a setting of distinct shapes based on the 'shape' variable!}
}
\value{
Makes a plot where colored dots (or other shapes) are overlayed onto a tSNE, PCA, ICA, ..., plot of choice.  var is the argument that sets how dots will be colored, and it can refer to either continuous (ex: "CD34" = gene expression) or discrete (ex: "ident" = clustering) data.
}
\description{
Show data overlayed on a tsne or pca or other reduction-type
}
\examples{
pbmc <- Seurat::pbmc_small
DBDimPlot("res.1", object = "pbmc")
DBDimPlot("ident", object = "pbmc", reduction.use = "pca", ellipse = TRUE, do.label = TRUE)
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
DBDimPlot("res.1")
DBDimPlot("ident", reduction.use = "pca", ellipse = TRUE, do.label = TRUE)
}
