% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{DBHeatmap}
\alias{DBHeatmap}
\title{Outputs a heatmap of the object}
\usage{
DBHeatmap(genes = NULL, object = DEFAULT, cells.use = NULL,
  cell.names.meta = NULL, data.type = "normalized",
  heatmap.colors = colorRampPalette(c("blue", "white", "red"))(50),
  cells.annotation = FALSE,
  cells.annotation.colors = rep(list(MYcolors),
  length(cells.annotation)), data.out = FALSE, highlight.genes = NULL,
  ...)
}
\arguments{
\item{genes}{c("gene1","gene2","gene3",...) = list of genes to put in the heatmap. REQUIRED.}

\item{object}{the Seurat or RNAseq object to draw from = name of object in "quotes". REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{cells.use}{Cells to include: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. *USE* in object@cell.names[*USE*])}

\item{cell.names.meta}{quoted "name" of a meta.data slot to use for naming the columns instead of the raw cell/sample names.  Default = just use the raw cell/sample names in the data slot.}

\item{data.type}{For obtaining the expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation), or "normalized.to.max" (= pulls normalized data, then divides by the maximum value)? DEFAULT = "normalized"}

\item{heatmap.colors}{the colors to use for the expression matrix. Default is a ramp from navy to white to red with 50 slices.}

\item{cells.annotation}{FALSE/ the name of a metadata slot containing how the cells/samples should be annotated. Default = FALSE.}

\item{cells.annotation.colors}{The colors to use for the annotations bar}

\item{data.out}{= If set to TRUE, the output will be a list of objects and and the script that would have been used for inputing into pheatmap.  NOTE: currently, a row#### arguement is added that should not be there.  I will remove this in a later release, but for now, know that you should remove this before trying to run the script.}

\item{highlight.genes}{= A list of genes whose names you would like to show, c("Gene1","Gene2","Gene3").  Only these genes will be named in the resulting heatmap.}

\item{...}{other arguments that will be passed to pheatmap. For scRNAseq data, I recommend running with cluster_cols=FALSE first because clustering of thousands of cells can tak a long time!  Common ones: main for adding a title, cluster_cols or cluster_rows = FALSE to turn clustering off for cells or genes, treeheight_row or treeheight_col = 0 to remove or a positive # for setting how large the trees on the side/top should be drawn}
}
\value{
Makes a heatmap where
}
\description{
This function is a wrapper for the pheatmap function of the pheatmap package.  Given a set of genes, it will spit out a basic heatmap from your RNAseq data, or just certain cells/samples.
}
\examples{
pbmc <- Seurat::pbmc_small
DBHeatmap(c("MS4A1","GNLY","CD3E","CD14","FCER1A",
            "FCGR3A","LYZ","PPBP","CD8A"),
          object = "pbmc",
          cells.annotation = "ident")
#For real data, you will have more cells than this truncated dataset,
# so I recommend turning off cell clustering when you are trying out tweaks by
# adding cluster_cols=FALSE
DBHeatmap(c("MS4A1","GNLY","CD3E","CD14","FCER1A",
            "FCGR3A","LYZ","PPBP","CD8A"),
          object = "pbmc",
          cells.annotation = "ident",
          cluster_cols=FALSE)

# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
DBHeatmap(c("MS4A1","GNLY","CD3E","CD14","FCER1A",
            "FCGR3A","LYZ","PPBP","CD8A"),
          cells.annotation = "ident")

}
