% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{DBPlot}
\alias{DBPlot}
\title{Plots continuous data on the y-axis, grouped in a settable way in the x-axis}
\usage{
DBPlot(var, object = DEFAULT, group.by, color.by, shape.by = "",
  cells.use = NULL, plots = c("jitter", "vlnplot"),
  data.type = "normalized", do.hover = FALSE, data.hover = var,
  color.panel = MYcolors, colors = c(1:length(color.panel)),
  theme = theme_classic(), main = NULL, sub = NULL, ylab = "make",
  y.breaks = NULL, min = NULL, max = NULL, xlab = NULL,
  labels = NULL, rotate.labels = TRUE, hline = NULL,
  hline.linetype = "dashed", hline.color = "black", jitter.size = 1,
  jitter.width = 0.2, jitter.color = "black", jitter.shapes = c(16,
  15, 17, 23, 25, 8), jitter.shape.legend.size = 3,
  boxplot.width = 0.2, boxplot.color = "black",
  boxplot.show.outliers = NA, boxplot.fill = T,
  reorder.x = 1:length(meta.levels(group.by, object)),
  title.legend = F, auto.title = T)
}
\arguments{
\item{var}{Target Variable = values, OR a gene or metadata in "quotes". REQUIRED.}

\item{object}{the Seurat or RNAseq Object = name of object in "quotes". REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{group.by}{"metadata" to use for separating values. REQUIRED.}

\item{color.by}{"metadata" to use for coloring. Affects boxplot and vlnplot fills. REQUIRED when using either.}

\item{shape.by}{"metadata" to use for setting the shape of jitter.  Default = just dots. Ignored if not a "quoted" metadata or "ident"}

\item{cells.use}{Cells to include: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. USE in object@cell.names[USE])}

\item{plots}{types of plots to include: possibilities = "jitter", "boxplot", "vlnplot". NOTE: The order matters, so use c("back","middle","front") when inputing multiple to put them in the order you want.}

\item{data.type}{For when plotting expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), or "relative" (= pulls normalized data, then uses the scale() funciton to produce a relative-to-mean representation)? Default = "normalized"}

\item{do.hover}{TRUE/FALSE. Default = F.  If set to true (and if there is a jitter plotted - the data it will work with) : object will be converted to a ggplotly object so that data about individual points will be displayed when you hover your cursor over them.  'data.hover' argument is used to determine what data to use.}

\item{data.hover}{list of variable names, c("meta1","gene1","meta2","gene2"). determines what data to show on hover when do.hover is set to TRUE.}

\item{color.panel}{the set of colors to draw from}

\item{colors}{indexes / or order, of colors from color.panel to actual use}

\item{theme}{Allows setting of a theme. Default = theme_classic when nothing is provided.}

\item{main}{plot title}

\item{sub}{plot subtitle}

\item{ylab}{y axis label, default is "var" or "var expression" if var is a gene.}

\item{y.breaks}{a list of breaks that should be used as major gridlines. c(break1,break2,break3,etc.) NOTE: The low and highs of this variable will override `min` and `max`.}

\item{min}{Use to set a custom minimum y-value to show.  Default = set based on the limits of the data in var.}

\item{max}{Use to set a custom maximum y-value to show.  Default = set based on the limits of the data in var.}

\item{xlab}{x axis label, default is blank (NULL)}

\item{labels}{c("label1","label2","label3"). = Names of the samples/groups if you wish to change them.  Default is the values in the group.by data. NOTE: you need to give at least as many labels as there are discrete values in the group.by data.}

\item{rotate.labels}{TRUE/FALSE. whether the labels should be rotated.  Default = FALSE = vertical labels.}

\item{hline}{y value(s) where a dashed horizontal line should go}

\item{hline.linetype}{Type of line.  Any ggplot linetype should work.  Defaults to "dashed"}

\item{hline.color}{color(s) of the horizontal line(s)}

\item{jitter.size}{the size of the jitter shapes.}

\item{jitter.width}{the width/spread of the jitter in the x direction}

\item{jitter.color}{the color of the jitter shapes}

\item{jitter.shapes}{the shapes to use.  Default is the first in the list, which corresponds to dots.}

\item{jitter.shape.legend.size}{Changes the size of the shape key in the legend.  Use a number.  OR, set to "none" to remove from the legend completely}

\item{boxplot.width}{the width/spread of the boxplot in the x direction}

\item{boxplot.color}{the color of the lines of the boxplot}

\item{boxplot.show.outliers}{whether outliers should by including in the boxplot. Default is FALSE when there is a jitter plotted, TRUE if no jitter.}

\item{boxplot.fill}{whether the boxplot should be filled in or not.}

\item{reorder.x}{sequence of numbers from 1:length(meta.levels(group.by)) for providing a new order for the samples.  Default = alphabetical then numerical.}

\item{title.legend}{whether to leave the title for the plot's legend}

\item{auto.title}{TRUE/FALSE = whether a default 'main' should be generated.}
}
\value{
Makes a plot where continuous data, grouped by sample, age, cluster, etc., on the x-axis is shown on the y-axis by a violin plot, boxplot, and/or dots (or other shapes)
}
\description{
Plots continuous data on the y-axis, grouped in a settable way in the x-axis
}
\examples{
pbmc <- Seurat::pbmc_small
DBPlot("CD14", object = "pbmc", group.by = "res.1", color.by = "res.1")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
DBPlot("CD14", group.by = "res.1", color.by = "res.1")
}
