% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demuxlet_Tools.R
\name{demux.SNP.summary}
\alias{demux.SNP.summary}
\title{Plots the number of SNPs sequenced per droplet}
\usage{
demux.SNP.summary(object = DEFAULT, group.by = "Lane",
  color.by = "Lane", plots = c("jitter", "boxplot"),
  boxplot.color = "grey30", ...)
}
\arguments{
\item{object}{the Seurat Object = name of object in "quotes". REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{group.by}{"metadata" to use for separating values. Default is "Lane".}

\item{color.by}{"metadata" to use for coloring. Default is "Lane".}

\item{plots}{Default = c("jitter","boxplot"). = the types of plots to include: possibilities = "jitter", "boxplot", "vlnplot". NOTE: The order matters, so use c("back","middle","front") when inputing multiple to put them in the order you want.}

\item{boxplot.color}{The color of the lines of the boxplot.}

\item{...}{extra arguments passed to DBPlot}
}
\value{
For a given Seurat object, plots a summary of how many SNPs were available to Demuxlet for making sample annotation calls.  Assumes that the number of SNPs are stored in a 'demux.N.SNPs' metadata slot, as would be the case if the Seurat object was created with the Import10XDemux function.
}
\description{
Plots the number of SNPs sequenced per droplet
}
\examples{
#Data required for an example would be rather large.  For an example, see the online vignette.
}
