% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.References.R
\docType{package}
\name{dittoSeq}
\alias{dittoSeq}
\title{dittoSeq}
\description{
This package was built to make the analysis and visualization of single-cell and bulk RNA-sequencing data accessible for both experienced and novice coders, and for color vision impaired individuals.
}
\details{
Includes many plotting functions (\code{\link{dittoPlot}}, \code{\link{dittoDimPlot}}, \code{\link{dittoBarPlot}}, \code{\link{dittoHeatmap}}, ...),
color adjustment functions (\code{\link{Simulate}}, \code{\link{Darken}}, \code{\link{Lighten}}),
and helper funtions (\code{\link{meta}}, \code{\link{gene}}, \code{\link{isMeta}}, \code{\link{getMetas}}, ...) to aid in making sense of single cell or bulk RNA sequencing data.
All included plotting functions produce a ggplot object (or pheatmap for dittoHeatmap) by default and can spit out a full plot with just a few arguments.
Many additional arguments are available for customization to generate complex publication-ready figures.

Default \code{\link{dittoColors}} are color blindness friendly and adapted from \href{https://www.nature.com/articles/nmeth.1618}{Wong B, "Points of view: Color blindness." Nature Methods, 2011.}

To report bugs, suggest new features, or ask for help, the best method is to create an issue on the github, \href{https://github.com/dtm2451/dittoSeq}{here}.
}
\author{
Daniel Bunis
}
