% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoPlot.R
\name{dittoSingleAxisDataGather}
\alias{dittoSingleAxisDataGather}
\title{Gathers data for single-axis plotting functions}
\usage{
dittoSingleAxisDataGather(main.var, object = DEFAULT,
  group.by = "Sample", color.by = group.by, extra.vars = NULL,
  cells.use = NULL, data.type = "normalized", do.hover = FALSE,
  data.hover = c(var, extra.vars))
}
\arguments{
\item{main.var}{Character representing the name of a metadata, gene, or "ident" to be plotted. OR character, numeric, or factor vector of length equal to the total # cells in the dataset.}

\item{object}{the Seurat or RNAseq Object = name of object in "quotes". REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{group.by}{"metadata" to use for separating values. REQUIRED.}

\item{color.by}{"metadata" to use for coloring. Affects boxplot and vlnplot fills. REQUIRED when using either.}

\item{extra.vars}{"metadata" to use for setting the shape of jitter.  Default = just dots. Ignored if not a "quoted" metadata or "ident"}

\item{cells.use}{Cells to include: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. USE in object@cell.names[USE])}

\item{data.type}{For when plotting expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation), or "normalized.to.max" (= pulls normalized data, then divides by the maximum value)? DEFAULT = "normalized"}

\item{do.hover}{TRUE/FALSE. Default = FALSE.  If set to true (and if there is a jitter plotted - the data it will work with) : object will be converted to a ggplotly object so that data about individual points will be displayed when you hover your cursor over them.  'data.hover' argument is used to determine what data to use.}

\item{data.hover}{list of variable names, c("meta1","gene1","meta2","gene2"). determines what data to show on hover when do.hover is set to TRUE.}
}
\value{
Generates Target_data and Others_data data.frames for use by single-axis DittoSeq plotters.
}
\description{
Gathers data for single-axis plotting functions
}
