% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{extDim}
\alias{extDim}
\title{Extracts the PC/tsne/IC/dim.reduction loadings of each cell/sample}
\usage{
extDim(reduction.use, dim = 1, object = DEFAULT)
}
\arguments{
\item{reduction.use}{quoted "reduction" name. = REQUIRED. Common types are "pca", "tsne", "ica", "cca", "cca.aligned".}

\item{dim}{#. which component to extract the loadings for (PC1 vs 2 vs 3)}

\item{object}{the Seurat or RNAseq object = REQUIRED, unless DEFAULT <- object has been run.}
}
\value{
Returns a list where [[1]]=embeddings = the loadings of cell/sample for a given dimensional reduction component, and [[2]] = name = the string name that should be used to refer to the components if going into a plot.
}
\description{
Extracts the PC/tsne/IC/dim.reduction loadings of each cell/sample
}
\examples{
pbmc <- Seurat::pbmc_small
out <- extDim("pca", 1, object = "pbmc")
out$embeddings #will be the cell loadings
out$name # for a "pca" reduction will typically be "PC#", which is the proper way of
         # calling principal component 1, PC1, in a plot axis.
#If you wanted to plot the PC1 loadings of all cells in the different clusters...

DBPlot(var = extDim("pca", 1, object = "pbmc")$embeddings,
       object = "pbmc",
       group.by = "res.1",
       color.by = "res.1",
       ylab = extDim("pca", 1, object = "pbmc")$name)

}
