% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{gen.key}
\alias{gen.key}
\title{Creates a 'key' for properly refering to a dimensional reduction component when the full name (ex: PCA1) woul dbe wrong.}
\usage{
gen.key(reduction.use)
}
\arguments{
\item{reduction.use}{quoted "reduction" name. = REQUIRED. Common types are "pca", "tsne", "ica", "cca", "cca.aligned".}
}
\value{
Returns a string name that should be used to refer to the components if going into a plot.
}
\description{
Creates a 'key' for properly refering to a dimensional reduction component when the full name (ex: PCA1) woul dbe wrong.
}
\examples{
gen.key("pca_5000_genes") # The function sees that pca is in the reduction name, so it outputs "PC"
#Output: "PC"
gen.key("tsne_5000_genes") # The function sees that pca is in the reduction name, so it outputs "PC"
#Output: "tSNE_"
}
