% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{get.genes}
\alias{get.genes}
\title{Returns the names of all genes of a target object.}
\usage{
get.genes(object = DEFAULT)
}
\arguments{
\item{object}{A target Seurat, SingleCellExperiment, or \linkS4class{RNAseq} object, OR the name of the target object in "quotes".}
}
\value{
A string vector, returns the names of all genes of the \code{object}.
}
\description{
Returns the names of all genes of a target object.
}
\examples{

pbmc <- Seurat::pbmc_small

# To see all genes of an object
get.genes(pbmc)

}
\seealso{
\code{\link{is.gene}} for returning all genes in an \code{object}

\code{\link{gene}} for obtaining the expression data of genes
}
