% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{get.metas}
\alias{get.metas}
\title{Returns the names of all meta.data slots of a target object.}
\usage{
get.metas(object = DEFAULT)
}
\arguments{
\item{object}{A target Seurat, SingleCellExperiment, or \linkS4class{RNAseq} object, OR the name of the target object in "quotes".}
}
\value{
A string vector, returns the names of all metadata slots of the \code{object}
}
\description{
Returns the names of all meta.data slots of a target object.
}
\examples{

pbmc <- Seurat::pbmc_small

# To see all metadata slots of an object
get.metas(pbmc)

}
\seealso{
\code{\link{is.meta}} for checking if certain metadata slots exist in an \code{object}

\code{\link{meta}} for obtaining the contants of metadata slots
}
