% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{get.reductions}
\alias{get.reductions}
\title{Returns the names of dimensional reduction slots within the object.}
\usage{
get.reductions(object = DEFAULT)
}
\arguments{
\item{object}{the SingleCellExperiment, Seurat, or RNAseq object = REQUIRED unless `DEFAULT <- "object"` has been run.}
}
\value{
Returns the names of all dimensional reduction slots within the object.
}
\description{
Returns the names of dimensional reduction slots within the object.
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
get.reductions(object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
get.reductions()
}
