% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-getters.R
\name{getGenes}
\alias{getGenes}
\title{Returns the names of all genes of a target object.}
\usage{
getGenes(object = DEFAULT)
}
\arguments{
\item{object}{A target Seurat, SingleCellExperiment, or \linkS4class{RNAseq} object, OR the name of the target object in "quotes".}
}
\value{
A string vector, returns the names of all genes of the \code{object}.
}
\description{
Returns the names of all genes of a target object.
}
\examples{

pbmc <- Seurat::pbmc_small

# To see all genes of an object
getGenes(pbmc)

}
\seealso{
\code{\link{isGene}} for returning all genes in an \code{object}

\code{\link{gene}} for obtaining the expression data of genes
}
\author{
Daniel Bunis
}
