% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{grab_legend}
\alias{grab_legend}
\title{Extract the legend from a ggplot}
\usage{
grab_legend(ggplot)
}
\arguments{
\item{ggplot}{The ggplot object that you would like to grab the legend from.  Can be any of the single plots from DittoSeq except DBHeatmap and demux.calls.summary}
}
\value{
The legend of a ggplot plot
}
\description{
This function will extract the legend of any ggplot.
}
\examples{
#Grab data
pbmc <- Seurat::pbmc_small

#Make a plot
DBDimPlot(pbmc)

#Extract the legend:
grab_legend(DBDimPlot(pbmc))

#Extract the legend of a stored plot:
ggplot <- DBDimPlot(pbmc)
grab_legend(ggplot)
}
