% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demuxlet_Tools.R
\name{Import10XDemux}
\alias{Import10XDemux}
\title{Imports 10X and demuxlet data and creates a Seurat object}
\usage{
Import10XDemux(CellRangerLocations, Lane.names = NA, Demuxlet.best,
  verbose = TRUE)
}
\arguments{
\item{CellRangerLocations}{list of locations where barcodes.tsv, genes.tsv, and matrix.mtx are located.}

\item{Lane.names}{how the lanes should be named (if you want to give them something different from the default = Lane1, Lane2, Lane3...)}

\item{Demuxlet.best}{the location of the .best output file from running of demuxlet}

\item{verbose}{whether to print messages about the stage of this process that is currently being run.}
}
\value{
Imports CellRanger 10X data from multiple lanes, merges those into a single Seurat object, then extracts the linked sample calls and demuxlet statistics from the demuxlet .best output file.  All relevant demuxlet information is stored as meta.data slots.
}
\description{
Imports 10X and demuxlet data and creates a Seurat object
}
\examples{
#Data required for an example would be rather large.  For an example, see the online vignette.
}
