% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demuxlet_Tools.R
\name{ImportDemux2Seurat}
\alias{ImportDemux2Seurat}
\title{Extracts Demuxlet information into a pre-made Seurat object}
\usage{
ImportDemux2Seurat(Seurat, Lane.info.meta = NULL, Lane.names = NA,
  Demuxlet.best, bypass.check = FALSE, verbose = TRUE)
}
\arguments{
\item{Seurat}{The Seurat object to add meta.datas to.}

\item{Lane.info.meta}{A meta.data}

\item{Lane.names}{how the lanes should be named (if you want to give them something different from the default = Lane1, Lane2, Lane3...)}

\item{Demuxlet.best}{the location of the .best output file from running of demuxlet}

\item{bypass.check}{for running this function anyway even when meta.data slots already around would be over-written.}

\item{verbose}{whether to print messages about the stage of this process that is currently being run.}
}
\value{
Imports Demuxlet sample calls and statistics into a Seurat object.  All relevant demuxlet information is stored as meta.data slots.  Works in 2 differnet ways:  If a Lane.info.meta is given, this metadata is copied into a "Lane" metadata becuase that is where my summarizer functions look by default.  If no Lane.info.meta is given, information about lane origin is assumed to be in the form of "-#" at the end of the cellnames; the form they would be in if 10X lanes were combined with `cellranger aggr`.
}
\description{
Extracts Demuxlet information into a pre-made Seurat object
}
\examples{
#Data required for an example would be rather large.
# For an example, see the online vignette at:
# https://github.com/dtm2451/DittoSeq/tree/master/Demuxlet-Vignette
}
