% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBulkRNAseq.R
\name{importEdgeR}
\alias{importEdgeR}
\title{Creates an RNAseq object from edgeR or limma-voom processed bulk RNAseq data objects.}
\usage{
importEdgeR(
  DGEList,
  normalized.data = NULL,
  normalization.fxn = "edgeR::cpm",
  counts = NULL,
  reductions = NULL
)
}
\arguments{
\item{DGEList}{a \link[edgeR]{DGEList-class} object, the main edgeR / Limma-Voom object for storing your data. REQUIRED.}

\item{normalized.data}{(optional) a matrix containing normalized data for all genes (rows) and samples (columns) in the DGEList.
When provided, this matrix is used directly and the \code{normalization.fxn} is not used to calculate the \code{data} slot from the \code{DGEList$counts} data.}

\item{normalization.fxn}{String which sets the counts data normalization function to use when calculating the \code{data} slot.
This is ignored when a matrix is provided to the \code{normalized.data} input.}

\item{counts}{(optional) a matrix containing raw counts data for all genes (rows) and samples (columns) in the DGEList.
When provided, this matrix is used, instead of the counts matrix embedded within the \code{DGEList}, for the RNAseq \code{counts} slot, but NOT for automatic calculation of the \code{data} slot.}

\item{reductions}{(optional) for advanced users, a named list containing individual dimensionality reductions perviously run.}
}
\value{
Outputs an \code{\linkS4class{RNAseq}} object.
}
\description{
The first step of visualization of edgeR- or Limma-Voom-analyzed bulk RNAseq data with dittoSeq is running this function to transform the data into the \linkS4class{RNAseq} structure that dittoSeq functions expect.
}
\details{
This function takes in a \code{DGEList} data structure and generates a dittoSeq \code{\linkS4class{RNAseq}} object with
raw \code{counts}, normalized \code{data}, \code{samples} names, and \code{meta.data} slots automatically pulled and filled in from the \code{DGEList}.
This effectively converts the data into the structure expected by all dittoSeq visualization and helper functions.

Because a common storage system for normalized counts within the DGEList structure is not set, for the \code{data} slot, this data is calculated with the edgeR \code{\link[edgeR]{cpm}} function.
Alternatively, this data can be directly supplied as a matrix to the \code{normalized.data} input, or
user's can provide the name of an alternative data normalization function to the \code{normalization.fxn} input to have the slot generated with this function instead (with default settings).

Metadata are pulled from the supplied \code{DGEList$samples} slot. Then "Sample" names and "Nreads" metadata are generated automatically if not already included.
}
\note{
dittoSeq is built to be a companion package for aiding analysis of RNAseq datasets of diverse structures.
It works quite well for this purpose, but is not meant to be used as the main tool for data-processing and normalization.
Inclusion of automatic normalization of bulk RNAseq data in import functions is provided for ease-of-use, but it is the responsibility of the user to provide properly normalized data
(to the `normalized.data` inputs of import functions) when the default settings for rlog (DESeq2) or cpm (edgeR/Limma-Voom) calculation are improper.
}
\examples{

# Generate mock RNAseq counts and a DGEList object from the mock data
# count tables from RNA-Seq data
counts.table <- matrix(rnbinom(n=1000, mu=100, size=1/0.5), ncol=10)
colnames(counts.table) <- paste0("Sample",1:10)
rownames(counts.table) <- paste0("Gene",1:100)
conditions <- factor(rep(1:2, each=5))
# object construction
library(edgeR)
DGE <- DGEList(counts=counts.table,group=conditions)
DGE <- calcNormFactors(DGE)
design <- model.matrix(~conditions)
DGE <- estimateDisp(DGE,design)

# Import
myRNA <- importEdgeR(DGE)
}
\seealso{
\code{\link[edgeR]{DGEList-class}}, and \code{\link[edgeR]{cpm}} for information about edgeR, the DGEList class, and the log2 counts per million calculation.

\code{\link{addMetaRNAseq}} for how to add additional metadata to an \code{RNAseq} object

\code{\link{addPrcomp}} and \code{\link{addDimReduction}} for how to add dimensionality reductions to an \code{\linkS4class{RNAseq}} object for use in
\code{\link{dittoDimPlot}} visualizations.
}
\author{
Daniel Bunis
}
