% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demuxlet_Tools.R
\name{ImportSeuratDemux}
\alias{ImportSeuratDemux}
\title{Imports demuxlet data into a (merged) Seurat object}
\usage{
ImportSeuratDemux(Seurats, Lane.names = NA, Demuxlet.best,
  verbose = TRUE)
}
\arguments{
\item{Seurats}{The set of Seurat objects that represent your dataset (each is assumed to represent individual 10X lanes).}

\item{Lane.names}{how the lanes should be named (if you want to give them something different from the default = Lane1, Lane2, Lane3...)}

\item{Demuxlet.best}{the location of the .best output file from running of demuxlet}

\item{verbose}{whether to print messages about the stage of this process that is currently being run.}
}
\value{
Imports Seurat objects, each expected to represent a single lane of droplet generation (or separate pool of samples for another sc-seq method, or sequencing lane, etc.), merges those into a single Seurat object, then extracts the linked sample calls and demuxlet statistics from the demuxlet .best output file.  All relevant demuxlet information is stored as meta.data slots.
}
\description{
Imports demuxlet data into a (merged) Seurat object
}
\examples{
#Data required for an example would be rather large.  For an example, see the online vignette.
}
