% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{is.gene}
\alias{is.gene}
\title{Tests if an input is the name of a gene recovered in the dataset.}
\usage{
is.gene(test, object = DEFAULT, value = FALSE)
}
\arguments{
\item{test}{"potential.gene.name" in quotes. REQUIRED.}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{value}{TRUE/FALSE. Default = FALSE.whether to return the gene names instead of TRUE/FALSE.}
}
\value{
Returns TRUE if there is a row in the objects' data slot named 'test'.
}
\description{
Tests if an input is the name of a gene recovered in the dataset.
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small

#For testing an individual query:
is.gene("CD14", object = "pbmc")
  # TRUE
is.gene("CD12345", pbmc)
  # FALSE, CD12345 is not a human gene.

# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped.
DEFAULT <- "pbmc"
is.gene("CD14")
  # TRUE

#The function works for sets of gene-queries as well
is.gene(c("CD14", "IL32", "CD3E", "CD12345"))
  #TRUE TRUE TRUE FALSE
# value input is especially useful in these cases.
is.gene(c("CD14", "IL32", "CD3E", "CD12345"), value = TRUE)
 #"CD14" "IL32" "CD3E"
}
