% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{is.meta}
\alias{is.meta}
\title{Tests if an input is the name of a meta.data slot.}
\usage{
is.meta(test, object = DEFAULT)
}
\arguments{
\item{test}{"potential.meta.data.name" in quotes. REQUIRED.}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless DEFAULT <- object has been run.}
}
\value{
Returns TRUE if there is a meta.data slot named 'test' (or for Seurat objects, if test = "ident", will also give TRUE because my meta() function knows how to handle meta("ident"))
}
\description{
Tests if an input is the name of a meta.data slot.
}
\examples{
pbmc <- Seurat::pbmc_small
is.meta("age", object = pbmc) #Returns FALSE because there is not a meta slot named "age"
is.meta("nUMI", object = pbmc) #Returns TRUE because there is a meta slot named "age"
get.metas(object = pbmc) #get.metas() will give a list of all the meta.data slot names.
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
is.meta("age")
get.metas()
}
