% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-get-and-set.R
\name{isBulk}
\alias{isBulk}
\title{Retrieve whether a SingleCellObject should be treated as single-cell versus bulk}
\usage{
isBulk(object)
}
\arguments{
\item{object}{A target SingleCellExperiment object

Alternatively, anything else, but then the result will always be \code{FALSE}}
}
\value{
Logical: whether the provided object would be treated as bulk data by dittoSeq
}
\description{
Retrieve whether a SingleCellObject should be treated as single-cell versus bulk
}
\examples{
example(importDittoBulk, echo = FALSE)
myRNA

isBulk(myRNA)
}
