% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{meta.levels}
\alias{meta.levels}
\title{Gives the distinct values of a meta.data slot (or ident)}
\usage{
meta.levels(meta, object = DEFAULT, cells.use = NULL)
}
\arguments{
\item{meta}{quoted "meta.data.slot" name = REQUIRED. the meta.data slot whose potential values should be retrieved.}

\item{object}{the Seurat or RNAseq object = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{cells.use}{String vector of cell/sample names to include, OR logical vector that is the same length as the number of cells in the object.
For the typically easier logical method, provide \code{USE} in \code{object@cell.names[USE]} OR \code{colnames(object)[USE]})}
}
\value{
Returns the distinct values of a metadata slot given to all cells/samples or for a subset of cells/samples.
(Alternatively, returns the distinct values of clustering if \code{meta = "ident"} and the object is a \code{Seurat} object).
}
\description{
Gives the distinct values of a meta.data slot (or ident)
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
meta.levels("RNA_snn_res.1", object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
meta.levels("RNA_snn_res.1")
}
