% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoDimPlot.R
\name{multi_dittoDimPlot}
\alias{multi_dittoDimPlot}
\title{Generates multiple dittoDimPlots arranged in a grid.}
\usage{
multi_dittoDimPlot(
  object,
  vars,
  legend.show = FALSE,
  ncol = NULL,
  nrow = NULL,
  axes.labels.show = FALSE,
  xlab = NA,
  ylab = NA,
  OUT.List = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to work with}

\item{vars}{c("var1","var2","var3",...). A list of vars from which to generate the separate plots}

\item{legend.show, xlab, ylab, ...}{other paramters passed to dittoDimPlot.}

\item{ncol, nrow}{Integer/NULL. How many columns or rows the plots should be arranged into}

\item{axes.labels.show}{Logical. Whether a axis labels should be shown. Ignored if xlab or ylab are set manually.}

\item{OUT.List}{Logical. (Default = FALSE) When set to \code{TRUE}, a list of the individual plots, named by the \code{vars} being shown in each, is output instead of the combined multi-plot.}
}
\value{
Given multiple 'var' parameters to \code{vars}, this function will output a dittoDimPlot for each one, arranged into a grid, with some slight tweaks to the defaults.
If \code{OUT.list} was set to TRUE, the list of individual plots, named by the \code{vars} being shown in each, is output instead of the combined multi-plot.
All parameters that can be adjusted in dittoDimPlot can be adjusted here, but the only parameter that can be adjusted between each is the \code{var}.
}
\description{
Generates multiple dittoDimPlots arranged in a grid.
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small

genes <- c("CD8A","CD3E","FCER1A","CD14","MS4A1")
multi_dittoDimPlot(pbmc, c(genes, "ident"))

}
\author{
Daniel Bunis
}
