% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_dittoDimPlotVaryCells.R
\name{multi_dittoDimPlotVaryCells}
\alias{multi_dittoDimPlotVaryCells}
\title{Generates multiple dittoDimPlots, each showing different cells, arranged into a grid.}
\usage{
multi_dittoDimPlotVaryCells(
  object,
  var,
  vary.cells.meta,
  vary.cells.levels = metaLevels(vary.cells.meta, object),
  assay = .default_assay(object),
  slot = .default_slot(object),
  adjustment = NULL,
  min = NULL,
  max = NULL,
  color.panel = dittoColors(),
  colors = seq_along(color.panel),
  show.titles = TRUE,
  show.allcells.plot = TRUE,
  allcells.main = "All Cells",
  show.legend.single = TRUE,
  show.legend.plots = FALSE,
  show.legend.allcells.plot = FALSE,
  nrow = NULL,
  ncol = NULL,
  OUT.List = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to work with}

\item{var}{String name of a "gene" or "metadata" (or "ident" for a Seurat \code{object}) to use for coloring the plots.
This is the data that will be displayed for each cell/sample.

Alternatively, can be a vector of same length as there are cells/samples in the \code{object}.
Discrete or continuous data both work. REQUIRED.}

\item{vary.cells.meta}{String name of a metadata that should be used for selecting which cells to show in each "varycells" plot. REQUIRED.}

\item{vary.cells.levels}{The values/groupings of the \code{vary.cells.meta} metadata that should get a plot.
Defaults to all levels of the metadata.}

\item{color.panel, colors, min, max, assay, slot, adjustment, ...}{additional parameters passed todittoDimPlot.
All parameters except for \code{cells.use}, \code{main}, and \code{legend.show} can be used.
A few suggestions: \code{reduction.use} for setting which dimensionality reduction space to use.
\code{xlab} and \code{ylab} can be set to \code{NULL} to remove the axes labels and provide extra room for the data.
\code{size} can be used to adjust the size of the dots.}

\item{show.titles}{Logical which sets whether titles should be added to the individual varycells plots}

\item{show.allcells.plot}{Logical which sets whether an additional plot showing all of the cells should be added.}

\item{allcells.main}{String which adjusts the title of the allcells plot. Default = "All Cells".  Set to \code{NULL} or \code{""} to remove.}

\item{show.legend.single}{Logical which sets whether to add a single legend as an additional plot. Default = TRUE.}

\item{show.legend.plots}{Logical which sets whether or not legends should be plotted in varycells plot. Default = FALSE.}

\item{show.legend.allcells.plot}{Logical which sets whether or a legend should be plotted in the allcells plot. Default = FALSE.}

\item{ncol, nrow}{Integers which set dimensions of the plot grid.}

\item{OUT.List}{Logical which controls whether the list of plots should be returned as a list instead of as a single grid arrangement of the plots.}
}
\value{
multiple dittoDimPlot \code{\link[ggplot2]{ggplot}}s either arranged in a grid OR as a list
}
\description{
Generates multiple dittoDimPlots, each showing different cells, arranged into a grid.
}
\details{
This function generates separate dittoDimPlots that show the same target data, but for distinct cells.
Which cells fall into which plot is controlled with the \code{vary.cells.meta} parameter.
When the quoted name of a metadata containing discrete groupings is given to \code{vary.cells.meta},
the function makes separate plots containing all cells/samples of each grouping.

If plots for only certain groupings of cells are wanted, names of the wanted groupings can be supplied to the \code{vary.cells.levels} input.

The function then appends a plot containing all groupings, titled as "All Cells" (unless otherwise changed with the \code{allcells.main} parameter),
as well as a single legend.  Either of these can be turned off with the \code{show.allcells.plot} and \code{show.legend.single} parameters.

Plots are either  output in a grid (default) with \code{ncol} columns and \code{nrow} rows,
or alternatively as a simple list of ggplots if \code{OUT.List} is set to \code{TRUE}.
In the list, the varycells plots will be named by the value of \code{vary.cells.meta} that they contain,
the allcells plot will be named "allcells" and the single legend will be named "legend".

Either continuous or discrete \code{var} data can be displayed.
\itemize{
\item For continuous data, the range of potential values is calculated at the start, and set, so that colors represent the same values accross all plots.
\item For discrete data, colors used in each plot are adjusted so that colors represent the same groupings accross all plots.
}
}
\examples{
pbmc <- Seurat::pbmc_small

multi_dittoDimPlotVaryCells(pbmc, "CD14", vary.cells.meta = "ident")

# This function can be used to quickly scan for differences in expression
#   within or accross clusters/cell types by providing a gene to 'var'
multi_dittoDimPlotVaryCells(pbmc, "CD14", vary.cells.meta = "ident")

# This function is also great for generating separate plots of each individual
#   element of a tsne/PCplot/umap. This can be useful to check for dispersion
#   of groups that might otherwise be hidden behind other cells/samples.
#   To do so, set 'var' and 'vary.cells.meta' the same.
multi_dittoDimPlotVaryCells(pbmc, "ident", vary.cells.meta = "ident")

# The function can also be used to quickly visualize how separate clustering
#   resolutions match up to each other, or perhaps how certain conditions of
#   cells disperse accross clusters.
multi_dittoDimPlotVaryCells(pbmc, "RNA_snn_res.0.8", vary.cells.meta = "ident")

# For an alternative method of viewing, and easily quantifying, how discrete
#   conditions of cells disperse accross clusters, see '?dittoBarPlot'

}
\seealso{
\code{\link{dittoDimPlot}} for the base DimPlot plotting function

\code{\link{multi_dittoDimPlot}} for plotting distinct \code{var}s accross plots instead of disctinct cells
}
\author{
Daniel Bunis
}
