% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{multiDBDimPlot}
\alias{multiDBDimPlot}
\title{Generates multiple DBDimPlots arranged into a grid.}
\usage{
multiDBDimPlot(vars, object = DEFAULT, show.legend = F, ncol = 3,
  nrow = NULL, add.title = T, axes.labels = F, ...)
}
\arguments{
\item{vars}{c("var1","var2","var3",...). REQUIRED. A list of vars from which to generate the separate plots}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{show.legend}{TRUE/FALSE. Whether or not you would like a legend to be plotted.  Default = FALSE}

\item{ncol}{#. How many plots should be arranged per row.  Default = 3.}

\item{nrow}{#/NULL. How many rows to arrange the plots into.  Default = NULL(/blank) --> becomes however many rows are needed to show all the data.}

\item{add.title}{TRUE/FALSE. Whether a title should be added.}

\item{axes.labels}{TRUE/FALSE. Whether a axis labels should be added}

\item{...}{other paramters that can be given to DBDimPlot function used in exactly the same way.}
}
\value{
Given multiple 'var' parameters, this function will output a DBDimPlot for each one, arranged into a grid.  All parameters that can be adjusted in DBDimPlot can be adjusted here, but the only parameter that can be adjusted between each is 'var'.
}
\description{
Generates multiple DBDimPlots arranged into a grid.
}
\examples{
pbmc <- Seurat::pbmc_small
genes <- c("CD8A","CD3E","FCER1A","CD14","MS4A1")
multiDBDimPlot(c(genes, "ident"), object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
multiDBDimPlot(c(genes, "ident"))
}
