% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{multiDBPlot}
\alias{multiDBPlot}
\title{Generates multiple DBPlots arranged into a grid.}
\usage{
multiDBPlot(vars, object = DEFAULT, group.by, color.by,
  show.legend = F, ncol = 3, nrow = NULL, add.title = T,
  ylab = F, ...)
}
\arguments{
\item{vars}{c("var1","var2","var3",...). REQUIRED. A list of vars from which to generate the separate plots}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless DEFAULT <- object has been run.}

\item{group.by}{"metadata" to use for separating values. REQUIRED.}

\item{color.by}{"metadata" to use for coloring. Affects boxplot and vlnplot fills. REQUIRED when using either.}

\item{show.legend}{TRUE/FALSE. Whether or not you would like a legend to be plotted.  Default = FALSE}

\item{ncol}{#. How many plots should be arranged per row.  Default = 3.}

\item{nrow}{#/NULL. How many rows to arrange the plots into.  Default = NULL(/blank) --> becomes however many rows are needed to show all the data.}

\item{add.title}{TRUE/FALSE. Whether a title should be added.}

\item{ylab}{TRUE/FALSE/"var". Whether a y-axis title should be added, TRUE/FALSE or "var".  If "var", then the var names will be used.  If TRUE, y-label for any gene vars will be "'var' expression"}

\item{...}{other paramters that can be given to DBPlot function used in exactly the same way.}
}
\value{
Given multiple 'var' parameters, this function will output a DBPlot for each one, arranged into a grid.  All parameters that can be adjusted in DBPlot can be adjusted here.
}
\description{
Generates multiple DBPlots arranged into a grid.
}
\examples{
pbmc <- Seurat::pbmc_small
genes <- c("CD8A","CD3E","FCER1A","CD14")
multiDBPlot(genes, object = "pbmc",
            group.by = "res.1", color.by = "res.1")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object
      # input can be skipped completely.
DEFAULT <- "pbmc"
multiDBPlot(genes,
            group.by = "res.1", color.by = "res.1")
#To make it output a grid that is 2x2, to add y-axis labels
# instead of titles, and to show legends...
multiDBPlot(genes,
            group.by = "res.1", color.by = "res.1",
            nrow = 2, ncol = 2,              #Make it 2x2
            add.title = FALSE, ylab = TRUE,  #Add y axis labels instead of titles
            show.legend = TRUE)              #Show legends
# To eliminate the "expression", change ylab = TRUE to ylab = "var"
multiDBPlot(genes,
            group.by = "res.1", color.by = "res.1",
            nrow = 2, ncol = 2,              #Make it 2x2
            add.title = FALSE, ylab = "var", #Add y axis labels without "expression"
            show.legend = TRUE)              #Show legends
}
