% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{new_multiDBDimPlot_vary_cells}
\alias{new_multiDBDimPlot_vary_cells}
\title{Generates multiple DBDimPlots, each showing different cells, arranged into a grid.}
\usage{
new_multiDBDimPlot_vary_cells(var, object = DEFAULT, cells.use.meta,
  cells.use.levels = meta.levels(cells.use.meta, object),
  all.cells.plot = T, show.legend = F, add.single.legend = T,
  ncol = 3, nrow = NULL, add.title = T, axes.labels = F,
  min = NULL, max = NULL, data.type = "normalized", OUT.List = F,
  main = "make", all.cells.main = "All Cells", ...)
}
\arguments{
\item{var}{name of a "gene", "meta.data", "ident", or a vector the length of the #cells or #samples in 'object'. REQUIRED. A var with which to color the separate plots.  Referencing discrete or continuous data are both allowed.}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{cells.use.meta}{The name of the meta.data that will be used for selecting cells. REQUIRED.}

\item{cells.use.levels}{The values/groupings of the cells.use.meta that you wish to show. NOTE: these will be put into the plot's main title in order to have them be identifiable.}

\item{all.cells.plot}{TRUE/FALSE, whether a plot showing all of the cells should be included at the end.}

\item{show.legend}{TRUE/FALSE. Whether or not you would like a legend to be plotted in every plot.  Default = FALSE}

\item{add.single.legend}{TRUE/FALSE, whether to add a single legend as an additional plot.}

\item{ncol}{#. How many plots should be arranged per row.  Default = 3.}

\item{nrow}{#/NULL. How many rows to arrange the plots into.  Default = NULL(/blank) --> becomes however many rows are needed to show all the data.}

\item{add.title}{TRUE/FALSE. Whether a title should be added.}

\item{axes.labels}{TRUE/FALSE. Whether a axis labels should be added.}

\item{min}{Works as in DimPlot, not required to be given, but sets the value associated with the minimum color.  All points with a lower value than this will get the same min.color.}

\item{max}{Works as in DimPlot, not required to be given, but sets the value associated with the maximum color.  All points with a higher value than this will get the same max.color.}

\item{data.type}{As with regular DBDimPlotting, For when plotting expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation), or "normalized.to.max" (= pulls normalized data, then divides by the maximum value)? DEFAULT = "normalized"}

\item{OUT.List}{TRUE/FALSE. (Default = FALSE) Whether the output should be a list of objects instead of the full plot.  Outputting as list allows direct input into plot_around().}

\item{main}{For setting a title for each individual plot.  Options: Leave as "make" and the titles will be which cells are in each plot.  "any.text" will set the same title for all plots.  NULL or "" will remove the titles.}

\item{all.cells.main}{For setting the title of the All.cells plot.  Default is All Cells.  Set to NULL or "" to remove.}

\item{...}{other paramters that can be given to DBDimPlot function used in the same way.}
}
\value{
A function for quickly making multiple DBDimPlots arranged in a grid, where instead of varying the 'var' displayed, what varies is the cells that are shown. Most parameters that can be adjusted in DBDimPlot can be adjusted here, but the only parameter that can be adjusted between each plot is which cells get displayed. NOTE: This function is incompatible with changing the 'colors' input. If you need to change the order of when certain colrs are chosen, change the order in color.panel. Also note: if 'var' given refers to continuous data, then the min and max will be calculated at the beginning in order to make the scale consistent accross all plots.  You can still set your own range, and this would also create a consistent scale.
}
\description{
Generates multiple DBDimPlots, each showing different cells, arranged into a grid.
}
\examples{
pbmc <- Seurat::pbmc_small
multiDBDimPlot_vary_cells("CD14", object = "pbmc", cells.use.meta = "ident")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
multiDBDimPlot_vary_cells("CD14", cells.use.meta = "ident")
}
