% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkRNAseq.R
\name{PCAcalc}
\alias{PCAcalc}
\title{Wrapper for running prcomp on an RNAseq object}
\usage{
PCAcalc(object = DEFAULT, genes.use = NULL, Ngenes = 2500,
  percent.samples = 75, name = "pca")
}
\arguments{
\item{object}{the RNAseq object = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{genes.use}{NULL or a vector of genes.  This will set the genes to be used by the prcomp PCA calculation.  NOTE: this list will not be used to populate the var.genes slot, but you can do that manually if you want to.}

\item{Ngenes}{#. How many genes to use for the PCA calculation. (This number will ultimately be the length of the var.genes slot)}

\item{percent.samples}{# between 0 and 100. The percent of samples within each condition that must express the gene in order for a gene to be included in the PCA calculation.}

\item{name}{"name". Example: "pca". The name to be given to this pca reduction slot. -> redution$'name'}
}
\value{
Outputs an RNAseq object with a new reductions$'name' slot.
}
\description{
This function will run prcomp PCA calculation on either a set of genes given by the genes.use slot, or on the Ngenes that have the highest coefficient of variation (CV=mean/sd) after a per experimental condition (extracted from the dds) expression filter is applied.
}
\examples{
# If bulkObject is the name of an RNAseq object in your workspace...
# PCAcalc("bulkObject", Ngenes = 2500, percent.samples = 75, name = "pca")
# Minimal input that does the same thing:
# PCAcalc("bulkObject")

}
