% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{remove_legend}
\alias{remove_legend}
\title{Remove the legend from a ggplot}
\usage{
remove_legend(ggplot)
}
\arguments{
\item{ggplot}{The ggplot object that you would like to eliminate the legend from.  Can be any of the single plots from DittoSeq except DBHeatmap and demux.calls.summary}
}
\value{
A ggplot plot with its legend removed.
}
\description{
This function will remove the legend of any ggplot.
}
\examples{
#Grab data
pbmc <- Seurat::pbmc_small

#Make a plot
DBDimPlot(pbmc)

#Remove the legend:
remove_legend(DBDimPlot(pbmc))

#Remove the legend of a stored plot:
ggplot <- DBDimPlot(pbmc)
remove_legend(ggplot)
}
