% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkRNAseq.R
\docType{class}
\name{RNAseq-class}
\alias{RNAseq-class}
\alias{Class}
\title{The RNAseq Class}
\description{
The RNAseq object stores data analyzed in DESeq2 in a structure similar to a Seurat-object.  This is the data structure required for DittoSeq plottign functions to access bulk RNAseq data.  All that is needed to create an RNAseq object is a DESeqDataSet output from the DESeq() function.
}
\section{Slots}{

\describe{
\item{\code{counts}}{a matrix. The raw genes x samples counts data. It is recommended, but not required, that one of these should be given when a new RNBAseq object is created.}

\item{\code{dds}}{a DESeqDataSet. The output of having run DESeq() on your data.}

\item{\code{data}}{a matrix. The regularized log correction of the counts data generated by a call to DESeq's rlog function.}

\item{\code{meta.data}}{a data.frame that contains meta-information about each sample. Autopopulated from the DESeq upon import, or added to manually afterward. Can be sample names, conditions, timepoints, Nreads (the number of reads).}

\item{\code{reductions}}{a list of dimensional reductions that have been run. Any type of reduction can technically be supported, but the only one included with my package is pca by prcomp. Embeddings are stored in the reductions$pca$x slot.}

\item{\code{var.genes}}{vector of genes with high coefficient of variation that passed an expression filter, and would therefore be included the default principal components analysis calculation.}

\item{\code{samples}}{a vector of names of the samples.}

\item{\code{exp.filter}}{a logical vector showing whether each gene passed the expression filter (default: at least 1 count in 75 percent of samples from each condition)}

\item{\code{CVs}}{a numeric vector showing the coefficient of variation (mean divided by sd) for each gene in the dataset.}

\item{\code{other}}{A great place to store any other data associated with your bulk experiment that does not fit elsewhere in the object.  Components of the list can be of any type.  Left empty by default, and is not altered or used by any of the DittoSeq functions.}
}}

