% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.References.R
\docType{data}
\name{RNAseq_mock}
\alias{RNAseq_mock}
\title{RNAseq_mock}
\format{
An object of class \code{RNAseq} of length 1.
}
\usage{
RNAseq_mock
}
\value{
An \code{\linkS4class{RNAseq}} object
}
\description{
An \code{\linkS4class{RNAseq}} object built from randomly generated data
}
\details{
The example code below was used to generate this object.
}
\examples{
# This is the code that was used to generate this object.
# RNAseq example
# # Generate mock RNAseq counts and a DESeq object from the mock data
# # count tables from RNA-Seq data
# set.seed(1)
# counts.table <- matrix(rnbinom(n=1000, mu=100, size=1/0.5), ncol=10)
# colnames(counts.table) <- paste0("Sample",1:10)
# rownames(counts.table) <- paste0("Gene",1:100)
# conditions <- factor(rep(1:2, each=5))
# # object construction
# library(DESeq2)
# dds <- DESeqDataSetFromMatrix(
#     counts.table, DataFrame(conditions), ~ conditions)
# dds <- DESeq(dds)
# # Import
# RNAseq_mock <- importDESeq2(dds, blind = FALSE)
# PCA <- prcomp(t(RNAseq_mock@data), center = TRUE, scale = TRUE)
# RNAseq_mock <- addDimReduction(
#     embeddings = PCA$x,
#     object = RNAseq_mock,
#     name = "pca",
#     key = "PC",
#     raw.object = PCA)

}
\seealso{
\code{\linkS4class{RNAseq}}
}
\author{
Daniel Bunis
}
\keyword{datasets}
