% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{S4_2string}
\alias{S4_2string}
\title{Turns an S4 object into it's name in string form}
\usage{
S4_2string(object = DEFAULT)
}
\arguments{
\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}
}
\value{
mainly for standardization within DittoSeq functions, outputs the string name and ensures objects can be handled in that form.
}
\description{
Turns an S4 object into it's name in string form
}
\examples{
pbmc <- Seurat::pbmc_small
S4_2string(pbmc)
S4_2string("pbmc")
}
