% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{var_OR_get_meta_or_gene}
\alias{var_OR_get_meta_or_gene}
\title{Outputs gene expression data, metadata data, or clustering data}
\usage{
var_OR_get_meta_or_gene(var, object = DEFAULT, data.type)
}
\arguments{
\item{var}{name of a metadata, gene, or "ident". = the data that should be grabbed}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{data.type}{For when extracting expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation), or "normalized.to.max" (= pulls normalized data, then divides by the maximum value)? DEFAULT = "normalized"}
}
\value{
determines what type of var is given, and outputs the gene expression data, metadata data, or clustering data.
}
\description{
Outputs gene expression data, metadata data, or clustering data
}
\examples{
pbmc <- Seurat::pbmc_small
var_OR_get_meta_or_gene("CD14", "pbmc", "normalized")
var_OR_get_meta_or_gene("ident", "pbmc")
var_OR_get_meta_or_gene("nUMI", "pbmc")
}
