% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{which_cells}
\alias{which_cells}
\title{Retrieves the list of cells.names to use}
\usage{
which_cells(cells.use, object = DEFAULT)
}
\arguments{
\item{cells.use}{either a logical or a list of names.}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}
}
\value{
Given a logical or a list of names (or NULL) will output the list of cells names.  For retrieval / standardization.
}
\description{
Retrieves the list of cells.names to use
}
\examples{
pbmc <- Seurat::pbmc_small
which_cells(meta("ident")=="0", "pbmc")
}
