% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{which_data}
\alias{which_data}
\title{Retrieves the proper slot of data}
\usage{
which_data(data.type, object = DEFAULT)
}
\arguments{
\item{data.type}{"raw", "normalized", or "scaled". REQUIRED. which type of data is requested}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}
}
\value{
Given "raw", "normalized", or "scaled", this function will output the proper slot of a seurat or RNAseq object.
}
\description{
Retrieves the proper slot of data
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
which_data("normalized", "pbmc")
}
