% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkRNAseq.R
\name{addDimReduction}
\alias{addDimReduction}
\title{Add any dimensionality reduction space to an RNAseq object}
\usage{
addDimReduction(embeddings, object = DEFAULT, name,
  key = .gen_key(name), raw.object = NULL)
}
\arguments{
\item{embeddings}{a numeric matrix containing the coordinates of all samples within the dimensionality reduction space.}

\item{object}{the \code{\linkS4class{RNAseq}} object. REQUIRED, unless \code{DEFAULT <- "object"} has been run.}

\item{name}{String name for the reduction slot. Example: "pca".
This will become the name of the slot, what should should be provided to the \code{reduction.use} input when making a \code{\link{dittoDimPlot}}.}

\item{key}{String, like "PC", which sets the default axes-label prefix when this reduction is used for making a \code{\link{dittoDimPlot}}.
If nothing is provided, a key will be automatically generated.}

\item{raw.object}{Optional, but recommended as this keeps the data all together: the output of the original calculation.}
}
\value{
Outputs an \code{\linkS4class{RNAseq}} object with a new \code{@reductions$'name'} slot.
}
\description{
Add any dimensionality reduction space to an RNAseq object
}
\examples{

# Generate mock RNAseq counts and a DESeq object from the mock data
# count tables from RNA-Seq data
counts.table <- matrix(rnbinom(n=1000, mu=100, size=1/0.5), ncol=10)
colnames(counts.table) <- paste0("Sample",1:10)
rownames(counts.table) <- paste0("Gene",1:100)
conditions <- factor(rep(1:2, each=5))

# object construction
library(DESeq2)
dds <- DESeqDataSetFromMatrix(
    counts.table, DataFrame(conditions), ~ conditions)
dds <- DESeq(dds)

# Import
myRNA <- importDESeq2(dds, blind = FALSE)

# Add a dimensionality reduction to myRNA
#   NOTE: This is typically not done with all genes in a dataset.
#   The inclusion of this example code is not an endorsement of a particular
#   method of PCA. Consult yourself, a bioinformatician, or literature for
#   tips on proper techniques.
PCA <- prcomp(t(myRNA@data), center = TRUE, scale = TRUE)
myRNA <- addDimReduction(
    embeddings = PCA$x,
    object = myRNA,
    name = "pca",
    key = "PC",
    raw.object = PCA)

### Add a batch metadata
# Option 1, all in the funciton:
myRNA <- addMetaRNAseq(
    value = rep(1:2, each = 5),
    name = "batch",
    object = myRNA)
# Option 2, create variable beforehand, and just provide that to value.
#     The name of the variable will be used to name the metadata slot.
batch <- rep(1:2, each = 5)
# Option 3, create variable beforehand, and provide just that using the
#     even further simplified setter function:
addMeta(myRNA) <- batch

# Visualize batch on a PCA plot
#   Note: For RNAseq objects, reduction.use defaults to "pca"
#   so just dittoDimPlot("Sample", myRNA, size = 3) would work the same!
dittoDimPlot("batch", myRNA, reduction.use = "pca", size = 3)

}
\seealso{
\code{\link{addPrcomp}} for a prcomp specific PCA import wrapper function

\code{\link{addMetaRNAseq}} for how to add metadata to an \code{RNAseq} object

\code{\link{importDESeq2}} for initial import of \code{RNAseq} data analyzed with DESeq2

\code{\link{dittoDimPlot}} for visualizing how samples group within this dimensionality reduction space after adding it to the object

\code{\linkS4class{RNAseq}} for learning more about the \code{RNAseq} object type
}
