% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoColors.R
\name{dittoColors}
\alias{dittoColors}
\title{Extracts the DittoSeq default colors}
\usage{
dittoColors(get.names = FALSE)
}
\arguments{
\item{get.names}{Logical, whether only the names of the default DittoSeq color panel should be returned instead}
}
\description{
Creates a colors vector of 24 colors in hexadecimal form,
a modification of the protanope and deuteranope friendly colors from Wong, B. Nature Methods, 2011.

The colors are:

1-7 = Suggested color panel from Wong, B. Nature Methods, 2011, minus black

8 = gray40

9-16 = 25\% darker versions of colors 1-8

17-24 = 25\% lighter versions of colors 1-8
}
\examples{
dittoColors()

#To retrieve names:
dittoColors(get.names = TRUE)
}
\author{
Daniel Bunis
}
