% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoDimPlot.R
\name{dittoDimPlot}
\alias{dittoDimPlot}
\title{Shows data overlayed on a tsne, pca, or similar type of plot}
\usage{
dittoDimPlot(var = "ident", object = DEFAULT, reduction.use = NA,
  dim.1 = 1, dim.2 = 2, theme = NA, size = 1, shape = 16,
  shapes = c(16, 15, 17, 23, 25, 8), legend.show = TRUE,
  legend.size = 5, legend.title = NULL, shape.legend.size = 5,
  shape.legend.title = NULL, data.type = "normalized", main = "make",
  sub = NULL, xlab = "make", ylab = "make", cells.use = NULL,
  show.others = TRUE, ellipse = FALSE, do.label = FALSE,
  label.size = 5, highlight.labels = TRUE, labels.repel = TRUE,
  rename.var.groups = NA, rename.shape.groups = NA,
  min.color = "#F0E442", max.color = "#0072B2", min = NULL,
  max = NULL, breaks = waiver(), color.panel = MYcolors,
  colors = 1:length(color.panel), do.letter = NA, do.hover = FALSE,
  data.hover = var, opacity = 1, data.out = FALSE,
  add.trajectories = NULL, trajectory.clusters,
  trajectory.arrow.size = 0.15)
}
\arguments{
\item{var}{Target Variable = either values or a metadata (in "quotes"), gene (in "quotes"), or "ident"}

\item{object}{the Seurat or RNAseq object to work on}

\item{reduction.use}{"pca", "tsne", "ica", etc.  Default = tsne for Seurat objects, and pca for RNAseq objects}

\item{dim.1}{The component number to use on the x-axis.  Default = 1}

\item{dim.2}{The component number to use on the y-axis.  Default = 2}

\item{theme}{Allows setting of a theme. Default = theme_bw when nothing is provided.}

\item{size}{Number. Size of data points.  Default = 1.}

\item{shape}{Number for setting shape OR name of metadata to use for setting shape}

\item{shapes}{the shapes to use.  Default is a list of 6.  There are more, but not many of the default ggplot options are great.  I recommend using colors for any variable with 7+ options.}

\item{legend.show}{Logical. Whether the legend should be displayed. Default = TRUE.}

\item{legend.size}{The size to increase the plotting of colors/letters legend shapes to (for discrete variable plotting)}

\item{legend.title}{For adding a title for the colors/letters legend.  It is set to NULL (off) by default.}

\item{shape.legend.size}{The size to increase the plotting of shapes legend shapes to.}

\item{shape.legend.title}{For adding a title for the shapes legend is a meta.data was given to 'shape' and multiple shapes were therefore used.  It is set to NULL (off) by default.}

\item{data.type}{For when plotting expression data, sets the data-type slot that will be obtained. See \link[DittoSeq]{gene} for options and details. DEFAULT = "normalized".}

\item{main}{plot title.  Default = "make", if left as "make", a title will be autogenerated based on the var input}

\item{sub}{plot subtitle}

\item{xlab}{label for y axes.  Default labels are generated if you do not give this a specific value.  To remove, set to NULL.}

\item{ylab}{label for y axes.  Default labels are generated if you do not give this a specific value.  To remove, set to NULL.}

\item{cells.use}{cells to show: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. *THIS*: object@cell.names[*THIS*])}

\item{show.others}{Logical. TRUE by default, whether other cells should be shown in the background}

\item{ellipse}{Logical. Whether the groups should be surrounded by an ellipse.}

\item{do.label}{Logical. Whether to add text labels at the center (median) of clusters for grouping vars}

\item{label.size}{Size of the the labels text}

\item{highlight.labels}{Logical. Whether the labels should have a box behind them}

\item{labels.repel}{Logical, that sets whether the labels' placements will be adjusted with \link{ggrepel} to avoid intersections between labels and plot bounds.
TRUE by default.}

\item{rename.var.groups}{new names for the identities of var.  Change to NULL to remove labeling altogether.}

\item{rename.shape.groups}{new names for the identities of shape groups.  Change to NULL to remove labeling altogether.}

\item{min.color}{color for lowest values of var/min}

\item{max.color}{color for highest values of var/max}

\item{min}{set the value associated with the minimum color.  All points with a lower value than this will get the same min.color.}

\item{max}{set the value associated with the maximum color.  All points with a higher value than this will get the same max.color.  Note: if your legend is not plotting, it's likely because min > max.}

\item{breaks}{Numeric vector. Sets the discrete values to show in the color-scale legend for continuous data.}

\item{color.panel}{a list of colors to be used for when plotting a discrete var.}

\item{colors}{indexes / order of colors from color.panel to use. USAGE= changing the order of how colors are linked to specific groups}

\item{do.letter}{Logical/NA. Whether letters should be added on top of the colored dots. For colorblindness compatibility.  NA by default, and if left that way, will be set to either TRUE or FALSE depending on the number of groups if a discrete var is given. For when there are lots of descrete variables, it can be hard to see by just color / shape.  NOTE: Lettering is incompatible with changing dots to shapes, so this will override a setting of distinct shapes based on the 'shape' variable!}

\item{do.hover}{Logical. Default = FALSE.  If set to true: object will be converted to a ggplotly object so that data about individual points will be displayed when you hover your cursor over them.  'data.hover' argument is used to determine what data to use.  NOTE: incompatible with lettering (due to a ggplotly incompatibility). Setting do.hover to TRUE will override a do.letter=TRUE or NA.}

\item{data.hover}{list of variable names, c("meta1","gene1","meta2","gene2"). determines what data to show on hover when do.hover is set to TRUE.}

\item{opacity}{Number between 0 and 1. Great for when you have MANY overlapping points, this sets how see-through the points should be; 1 = not at all; 0 = invisible. Default = 1.}

\item{data.out}{Whether just the plot should be output, or a list with the plot and Target_data and Others_data dataframes.  Note: plotly output is turned off in this setting, but hover.data is still calculated.}

\item{add.trajectories}{List vectors representing trajectory paths from start-cluster to end-cluster where vector contents are the names of clusters provided in the \code{trajectory.clusters} input.
If Slingshot package was used for trajectory analysis, you can use \code{SlingshotDataSet(SCE_with_slingshot)$lineages}.}

\item{trajectory.clusters}{String name of metadata containing the clusters that were used for generating trajectories.  Names of clusters must be the same as the contents of \code{add.trajectories} vectors.}

\item{trajectory.arrow.size}{Number representing the size of trajectory arrows in inches.  Default = 0.15.}
}
\value{
Makes a plot where colored dots (or other shapes) are overlayed onto a tSNE, PCA, CCA, ..., plot of choice.  var is the argument that sets how dots will be colored, and it can refer to either continuous (ex: "CD34" = gene expression) or discrete (ex: "ident" = clustering) data.
}
\description{
Shows data overlayed on a tsne, pca, or similar type of plot
}
\details{
If \code{data=TRUE}, a list containing the plot (p), a data.table containing the underlying data for target cells (Target_data), and a data.table containing the underlying data for non-target cells (Others_data).
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
dittoDimPlot("RNA_snn_res.1", object = "pbmc")
#To show currently set clustering, you can use "ident".
#To change the dimensional reduction type, use reduction.use.
#MANY other tweaks are possible!
dittoDimPlot("ident", object = "pbmc", reduction.use = "pca", ellipse = TRUE, do.label = TRUE)

# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
dittoDimPlot("RNA_snn_res.1")
dittoDimPlot("ident", reduction.use = "pca", ellipse = TRUE, do.label = TRUE)
}
