% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoHeatmap.R
\name{dittoHeatmap}
\alias{dittoHeatmap}
\title{Outputs a heatmap of given genes}
\usage{
dittoHeatmap(genes = NULL, object = DEFAULT, cells.use = NULL,
  main = NA, cell.names.meta = NULL, data.type = "normalized",
  heatmap.colors = colorRampPalette(c("blue", "white", "red"))(50),
  scaled.to.max = FALSE,
  heatmap.colors.max.scaled = colorRampPalette(c("white", "red"))(25),
  col.annotation.metas = NULL, annotation.colors = rep(list(MYcolors),
  length(col.annotation.metas)), data.out = FALSE,
  highlight.genes = NULL, show.colnames = TRUE, show.rownames = TRUE,
  ...)
}
\arguments{
\item{genes}{c("gene1","gene2","gene3",...) = list of genes to put in the heatmap. REQUIRED.}

\item{object}{String name of the object to draw from.  Alternatively, the \code{\link[Seurat]{Seurat}},  or \code{\linkS4class{RNAseq}}, or \code{\link[SingleCellExperiment]{SingleCellExperiment}} object itself. REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{cells.use}{String vector of cell/sample names to include, OR logical vector that is the same length as the number of cells in the object.}

\item{main}{String that sets the title for the heatmap.}

\item{cell.names.meta}{quoted "name" of a meta.data slot to use for naming the columns instead of the raw cell/sample names.}

\item{data.type}{String. Options are "normalized" (data slot, default), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects). Note: scaling is performed on the data matrix by default.}

\item{heatmap.colors}{the colors to use within the heatmap.
Default is a ramp from navy to white to red with 50 slices.}

\item{heatmap.colors.max.scaled}{the colors to use within the heatmap when \code{scaled.to.max} is set to \code{TRUE}.
Default is a ramp from white to red with 25 slices.}

\item{col.annotation.metas}{String name of a metadata slot containing how the cells/samples should be annotated. Default = \code{NULL}.}

\item{annotation.colors}{List of String (color) vectors where each vector is the set of colors for an individual annotation bar.}

\item{data.out}{= If set to TRUE, the output will be a list containing \code{args},
a list of arguments passed to \code{pheatmap},
and \code{call} containing how \code{pheatmap} would have been called.}

\item{highlight.genes}{= A list of genes whose names you would like to show, c("Gene1","Gene2","Gene3").  Only these genes will be named in the resulting heatmap.}

\item{show.rownames, show.colnames}{Logical which sets whether rownames or colnames will be shown.
Note: if gene names are provided to \code{highlight.genes}, the \code{show.colnames} parameter is ignored.}

\item{...}{other arguments that will be passed to \code{\link[pheatmap]{pheatmap}}.
For scRNAseq data, I recommend running with \code{cluster_cols=FALSE} first because clustering of thousands of cells can tak a long time!
Common options are \code{main} for adding a title,\code{cluster_cols} or \code{cluster_rows} for controlling clustering, \code{treeheight_row} or \code{treeheight_col} for setting how large the trees on the side/top should be drawn.}

\item{scale.to.max}{Logical which sets whether expression shoud be scaled between [0, 1].
This is recommended for single-cell datasets as they are generally enriched in 0s.}
}
\value{
A \code{pheatmap} object.
}
\description{
This function is a wrapper for the \code{\link[pheatmap]{pheatmap}} function of the \code{pheatmap} package.
Given a set of genes, cells, and annotation metadata names, it will retrieve the expression data for those genes and cells, and the annotation data for those cells,
then it will utilize the \code{pheatmap} function to make a heatmap.
}
\details{
Notes on adding row annotations:

When \code{annotation_row} parameter in \code{\link[pheatmap]{pheatmap}} is used to add row annotations,
extra color vectors must be added to the \code{annotation.colors} input to accomodate these.
The extra vectors should be named with the annotation bar titles, should come after vectors provided for column annotations.
If the vectors are not named, the \code{pheatmap}'s default colors will be used instead.
}
\seealso{
\code{\link[pheatmap]{pheatmap}}, for how to add additional heatmap tweaks.
library(Seurat)
pbmc <- Seurat::pbmc_small
dittoHeatmap(c("MS4A1","GNLY","CD3E","CD14","FCER1A",
    "FCGR3A","LYZ","PPBP","CD8A"),
    object = "pbmc",
    col.annotation.metas = "ident")
#For real data, you will have more cells than this truncated dataset,
# so I recommend turning off cell clustering when you are trying out tweaks by
# adding cluster_cols=FALSE
dittoHeatmap(c("MS4A1","GNLY","CD3E","CD14","FCER1A",
    "FCGR3A","LYZ","PPBP","CD8A"),
    object = "pbmc",
    col.annotation.metas = "ident",
    cluster_cols=FALSE)

# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
dittoHeatmap(c("MS4A1","GNLY","CD3E","CD14","FCER1A",
    "FCGR3A","LYZ","PPBP","CD8A"),
    col.annotation.metas = "ident")
}
