% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoScatterPlot.R
\name{dittoScatterPlot}
\alias{dittoScatterPlot}
\title{Show RNAseq data overlayed on a scatter plot}
\usage{
dittoScatterPlot(x.var, y.var, overlay.color.var = NULL,
  overlay.shape.var = NULL, object = DEFAULT, cells.use = NULL,
  show.others = FALSE, color.panel = MYcolors, colors = 1:24,
  data.type.x = "normalized", data.type.y = "normalized",
  data.type.color = "normalized", do.hover = FALSE,
  data.hover = NULL, data.type.hover = "normalized", shape = 16,
  shapes = c(16, 15, 17, 23, 25, 8), size = 1, opacity = 1,
  rename.color.groups = NA, rename.shape.groups = NA,
  legend.show = TRUE, legend.color.title = overlay.color.var,
  legend.color.size = 5, legend.shape.title = overlay.shape.var,
  legend.shape.size = 5, min.color = "#F0E442",
  max.color = "#0072B2", min = NULL, max = NULL, breaks = waiver(),
  xlab = x.var, ylab = y.var, main = NULL, sub = NULL,
  theme = theme_bw(), data.out = FALSE)
}
\arguments{
\item{x.var}{Variable for setting x-axis position of cells/samples.  Can be the name of a gene, meta-data, or "ident" for clusters of a Seurat object.  Alternatively, can be a numeric of length equal to the total number of cells/samples in object.}

\item{y.var}{Variable for setting y-axis position of cells/samples.  Can be the name of a gene, meta-data, or "ident" for clusters of a Seurat object.  Alternatively, can be a numeric of length equal to the total number of cells/samples in object.}

\item{overlay.color.var}{Variable for setting the color of cells/samples in the plot.  Can be the name of a gene or meta-data.  Alternatively, can be "ident" for clusters of a Seurat object.  Alternatively, can be a numeric of length equal to the total number of cells/samples in object.}

\item{overlay.shape.var}{Variable for setting the shape of cells/samples in the plot.  Note: must be discrete.  Can be the name of a gene or meta-data.  Alternatively, can be "ident" for clusters of a Seurat object.  Alternatively, can be a numeric of length equal to the total number of cells/samples in object.}

\item{object}{the Seurat, SingleCellExperiment, or RNAseq object to work on.}

\item{cells.use}{cells to show: either in the form of a character list of names, or a logical that is the same length as the number of cells in the object (a.k.a. *THIS*: object@cell.names[*THIS*])}

\item{show.others}{TRUE/FALSE. TRUE by default, whether other cells should be shown in the background}

\item{color.panel}{a list of colors to be used for when plotting a discrete var.}

\item{colors}{indexes / order of colors from color.panel to use. USAGE= changing the order of how colors are linked to specific groups}

\item{data.type.x}{For when plotting expression data, sets the data-type slot that will be obtained. See \link[DittoSeq]{gene}. DEFAULT = "normalized"}

\item{data.type.y}{For when plotting expression data, sets the data-type slot that will be obtained. See \link[DittoSeq]{gene}. DEFAULT = "normalized"}

\item{data.type.color}{For when plotting expression data, sets the data-type slot that will be obtained. See \link[DittoSeq]{gene}. DEFAULT = "normalized"}

\item{do.hover}{TRUE/FALSE. Default = FALSE.  If set to true: object will be converted to a ggplotly object so that data about individual points will be displayed when you hover your cursor over them.  'data.hover' argument is used to determine what data to use.  NOTE: incompatible with lettering (due to a ggplotly incompatibility). Setting do.hover to TRUE will override a do.letter=TRUE or NA.}

\item{data.hover}{list of variable names, c("meta1","gene1","meta2","gene2"). determines what data to show on hover when do.hover is set to TRUE.}

\item{data.type.hover}{For when plotting expression data, sets the data-type slot that will be obtained. See \link[DittoSeq]{gene}. DEFAULT = "normalized" For when plotting expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation), or "normalized.to.max" (= pulls normalized data, then divides by the maximum value)? DEFAULT = "normalized"}

\item{shape}{Number for setting the shape}

\item{shapes}{If multiple shapes are needed, this sets the sahpes to pull from.  Default is a list of 6, \code{c(16,15,17,23,25,8)}.  There are more, but not many of the default ggplot options are great.  Unfortunately, shape is also harder to see when points are on top of each other.  For these reasons, even as a color blind person myself, I recommend use of colors for variables with 7+ options.}

\item{size}{Number. Size of data points.  Default = 1.}

\item{opacity}{Number between 0 and 1. Great for when you have MANY overlapping points, this sets how see-through the points should be; 1 = not at all; 0 = invisible. Default = 1.}

\item{rename.color.groups}{Character vector containing new names for the identities of the color overlay.}

\item{rename.shape.groups}{Character vector containing new names for the identities of the shape overlay.}

\item{legend.show}{TRUE/FALSE. Whether the legend should be displayed. Default = TRUE.}

\item{legend.color.title}{For adding a title to the colors legend.  Set to \code{NULL} to turn off.}

\item{legend.color.size}{Override for legend size of the color variable.}

\item{legend.shape.title}{For adding a title for the colors legend.  Set to \code{NULL} to turn off}

\item{legend.shape.size}{Override for legend size of the shape variable.}

\item{min.color}{color for lowest values shown of the color overlay.  Default is \code{MYcolors[4]}, a yellow.}

\item{max.color}{color for highest values shown of the color overlay.  Default is \code{MYcolors[5]}, a dark blue.}

\item{min}{set the value associated with the minimum color.  All points with a lower value than this will get the same min.color.}

\item{max}{set the value associated with the maximum color.  All points with a higher value than this will get the same max.color.  Note: if your legend is not plotting, it may be because min > max.}

\item{breaks}{Numeric vector. Sets the discrete values to show in the color-scale legend for continuous data.}

\item{xlab}{label for y axes.  Default = \code{x.var}. To remove, set to NULL.}

\item{ylab}{label for y axes.  Default = \code{x.var}. To remove, set to NULL.}

\item{main}{plot title.  Default = \code{NULL}}

\item{sub}{plot subtitle.  Default = \code{NULL}}

\item{theme}{Allows setting of a theme. Default = theme_bw when nothing is provided.}

\item{data.out}{Whether just the plot should be output, or a list with the plot and Target_data and Others_data dataframes.  Note: plotly output is turned off in this setting, but hover.data is still calculated.}
}
\value{
Makes a plot where colored dots and/or shapes representing individual cells/samples are overlayed onto a scatterplot where x and y can be gene expression (or any numeric metadata) of those cells/samples.
}
\description{
Show RNAseq data overlayed on a scatter plot
}
