% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{gene}
\alias{gene}
\title{Returns the values of a gene for all cells/samples}
\usage{
gene(gene, object = DEFAULT, data.type = "normalized")
}
\arguments{
\item{gene}{quoted "gene" name = REQUIRED. the gene whose expression data should be retrieved.}

\item{object}{"name" of the Seurat or RNAseq object = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{data.type}{Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), or "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation)? Default = "normalized"}
}
\value{
Returns the values of a meta.data slot, or the ident (clustering) slot if "ident" was given and the object is a Seurat object.
}
\description{
Returns the values of a gene for all cells/samples
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
gene("CD14", object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
gene("CD14")
}
