% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{gene}
\alias{gene}
\title{Returns the values of a gene for all cells/samples}
\usage{
gene(gene, object = DEFAULT)
}
\arguments{
\item{gene}{quoted "gene" name = REQUIRED. the gene whose expression data should be retrieved.}

\item{object}{the Seurat or RNAseq object = REQUIRED, unless DEFAULT <- object has been run.}
}
\value{
Returns the values of a meta.data slot, or the ident (clustering) slot if "ident" was given and the object is a Seurat object.
}
\description{
Returns the values of a gene for all cells/samples
}
\examples{
pbmc <- Seurat::pbmc_small
gene("CD14", object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
gene("CD14")
}
