% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-getters.R
\name{gene}
\alias{gene}
\title{Returns the expression values of a gene for all cells/samples}
\usage{
gene(
  gene,
  object,
  assay = .default_assay(object),
  slot = .default_slot(object),
  adjustment = NULL,
  adj.fxn = NULL,
  swap.rownames = NULL
)
}
\arguments{
\item{gene}{quoted "gene" name = REQUIRED. the gene whose expression data should be retrieved.}

\item{object}{A Seurat, SingleCellExperiment, or SummarizedExperiment object.}

\item{assay, slot}{single strings or integer that set which data to use.
Seurat and SingleCellExperiments deal with these differently, so be sure to check the documentation for whichever object you are using.
When not provided, these typical defaults for the provided \code{object} class are used:
\itemize{
\item{SingleCellExperiment (single-cell or bulk data): \code{assay} = "logcounts", "normcounts", "counts", or the first element of assays(object), \code{slot} not used}
\item{Seurat-v3: \code{assay} = DefaultAssay(object), \code{slot} = "data"}
\item{Seurat-v2: \code{assay} not used, \code{slot} = "data"}
}}

\item{adjustment}{Should expression data be used directly (default) or should it be adjusted to be
\itemize{
\item{"z-score": scaled with the scale() function to produce a relative-to-mean z-score representation}
\item{"relative.to.max": divided by the maximum expression value to give percent of max values between [0,1]}
}}

\item{adj.fxn}{A function which takes a vector (of metadata values) and returns a vector of the same length.

For example, \code{function(x) \{log2(x)\}} or \code{as.factor}}

\item{swap.rownames}{String. For SummarizeedExperiment or SingleCellExperiment objects, the column name of rowData(object) to be used to identify features instead of rownames(object).}
}
\value{
Returns the expression values of a gene for all cells/samples.
}
\description{
Returns the expression values of a gene for all cells/samples
}
\examples{
example(importDittoBulk, echo = FALSE)
gene("gene1", object = myRNA, assay = "counts")

# z-scored
gene("gene1", object = myRNA, assay = "counts",
    adjustment = "z-score")

# Log2'd
gene("gene1", object = myRNA, assay = "counts",
    adj.fxn = function(x) \{log2(x)\})

# To see expression of the gene for the default assay that dittoSeq would use
# leave out the assay input
# (For this object, the default assay is the logcounts assay)
gene("gene1", myRNA)

# Seurat (raw counts)
if (!requireNamespace("Seurat")) {
    gene("CD14", object = Seurat::pbmc, assay = "RNA", slot = "counts")
}

}
\author{
Daniel Bunis
}
