% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{get.metas}
\alias{get.metas}
\title{Returns the names of all meta.data slots in the object.}
\usage{
get.metas(object = DEFAULT)
}
\arguments{
\item{object}{the Seurat or RNAseq object = REQUIRED, unless `DEFAULT <- "object"` has been run.}
}
\value{
Returns the names of all meta.data slots in the object.
}
\description{
Returns the names of all meta.data slots in the object.
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
get.metas(object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
get.metas()
}
