% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-getters.R
\name{get.metas}
\alias{get.metas}
\title{Returns the names of all meta.data slots of a target object.}
\usage{
get.metas(object = DEFAULT, names.only = TRUE)
}
\arguments{
\item{object}{A target Seurat, SingleCellExperiment, or \linkS4class{RNAseq} object, OR the name of the target object in "quotes".}

\item{names.only}{Logical, \code{TRUE} by default, which sets whether just the names should be output versus the entire metadata dataframe.}
}
\value{
A string vector of the names of all metadata slots of the \code{object}, or alternatively the entire dataframe of metadatas if \code{names.only} is set to \code{FALSE}
}
\description{
Returns the names of all meta.data slots of a target object.
}
\examples{

pbmc <- Seurat::pbmc_small

# To see all metadata slots of an object
get.metas(pbmc)

# To retrieve the entire metadata matrix
get.metas(pbmc, names.only = FALSE)

}
\seealso{
\code{\link{is.meta}} for checking if certain metadata slots exist in an \code{object}

\code{\link{meta}} for obtaining the contants of metadata slots
}
