% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.reductions.R
\name{get.reductions}
\alias{get.reductions}
\title{Returns the names of all dimensionality reduction slots of a target object.}
\usage{
get.reductions(object = DEFAULT)
}
\arguments{
\item{object}{A target Seurat, SingleCellExperiment, or \linkS4class{RNAseq} object, OR the name of the target object in "quotes".}
}
\value{
A string vector of the names of all dimensionality reduction slots of the \code{object}.
These represent the options for the \code{reduction.use} input of \code{\link{dittoDimPlot}}.
}
\description{
Returns the names of all dimensionality reduction slots of a target object.
}
\examples{

pbmc <- Seurat::pbmc_small

# To see all metadata slots of an object
get.reductions(pbmc)

}
\author{
Daniel Bunis
}
