% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demuxlet_Tools.R
\name{importDemux2Seurat}
\alias{importDemux2Seurat}
\title{Extracts Demuxlet information into a pre-made Seurat object}
\usage{
importDemux2Seurat(
  Seurat,
  lane.meta = NULL,
  lane.names = NA,
  Demuxlet.best,
  bypass.check = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Seurat}{A pre-made Seurat object to add demuxlet information to.}

\item{lane.meta}{A meta.data}

\item{lane.names}{how the lanes should be named (if you want to give them something different from the default = Lane1, Lane2, Lane3...)}

\item{Demuxlet.best}{String pointing to the location of the .best output file from running of demuxlet OR a data.frame representing already imported .best matrix.

Alternatively, can be a String vector representing the .best output locations of separate demuxlet runs which each correspond to 10X lanes that were combined together with cellranger aggr.  See the cellranger note below.}

\item{bypass.check}{for running this function anyway even when meta.data slots already around would be over-written.}

\item{verbose}{whether to print messages about the stage of this process that is currently being run.}
}
\value{
A Seurat object with sample calls and relevant statistics added as metadata.
}
\description{
Extracts Demuxlet information into a pre-made Seurat object
}
\details{
The function takes in a previously generated Seurat object.
It also takes in demuxlet information either in the form of
1: the location of a single demuxlet.best out file,
2: the locations of multiple demuxlet.best output files,
or 3: a user-constructed data.frame created by reading in a demuxlet.best file.

If a metadata slot name is provided to \code{lane.meta}, information in that metadata slot is copied into a metadata slot called "Lane".
Alternatively, if \code{lane.meta} is left as \code{NULL}, separate lanes are assumed to be marked by distinct values of "-#", as is the typical output of the 10X cellranger count & aggr pipeline.
In these situations, the \code{lane.names} input can be used to set specific names for each lane. "Lane1", "Lane2", "Lane3", etc, are used b y default.

Barcodes (minus any "***_" often added by Seurat when objects are merged) are then matched between the \code{Seurat} and the \code{Demuxlet.best} dataframes.

Note: "-#" information added by cellranger functions is not removed.
Doing so would cause cells, from separate 10X wells, which ended up with similar barcodes to become indistinguishable.
In demuxlet itself, ignorance of lane information leads to artificial doublet calls.
In \code{importDemux2Seurat}, ignorance of lane information can lead to import of improper demuxlet annotations.
For this reason, \code{importDemux2Seurat} checks for whether such artificial duplicates likely happened.
See the recommended cellranger/demuxlet pipeline below for specific suggestions for how to use this function with multi-well 10X data.

Lane information and demuxlet calls and statistics are then imported into the \code{Seurat} as metadata:
\itemize{
\item Lane = guided by \code{lane.meta} import input, represents of separate droblet-generation lane, pool, sequencing lane, etc.
\item Sample = The sample call, parsed from the BEST column
\item demux.doublet.call = whether the sample was a singlet (SNG), doublet (DBL), or ambiguious (AMB), parsed from the BEST column
\item demux.RD.TOTL = RD.TOTL column
\item demux.RD.PASS = RD.PASS column
\item demux.RD.UNIQ = RD.UNIQ column
\item demux.N.SNP = N.SNP column
\item demux.PRB.DBL = PRB.DBL column
\item demux.barcode.dup = (Only generated when TRUEs will exist) whether a cell's barcode in the Demuxlet.best refered to only 1 cell in the Seurat object.
(When TRUE, indicates that cells from distinct lanes were interpretted together by demuxlet.
These will often be mistakenly called as doublets.)
}

Finally, a summary of the results including mean number of SNPs and percentages of singlets and doublets is output unless \code{verbose} is set to \code{FALSE}.
}
\note{
For use of multi-well 10X data + Demuxlet + this function:
10X recommends running cellranger counts individually for each well/lane.
This leads to creation of separate genes x cells counts matrices for each lane.
*Demuxlet should also be run separately for each lane in order to minimize the informatic generation of artificial doublets.
Afterwards, there are two common methods of importing/merging such multi-well 10X data into Seurat.
Technical differences: Both options will alter the cell barcode names in a way that makes them unique across lanes, but how they do so is different.
Technical issue: Neither method adjusts the bacode names that are embedded within the BAM files which a user must supply to Demuxlet,
so that data needs to be modified in a proper way in order to make the Seurat cellnames and demuxlet BARCODEs match.

\code{importDemux2Seurat} can work with either.
\itemize{
\item Option 1: merging matrices of all lanes with cellranger aggr before Seurat import.
Barcode uniquification method: A "-1", "-2", "-3", ... "-#" is appended to the end of all barcode names.
The number is incremented for each succesive lane. (Note: lane-numbers depend on the order in which they were supplied to cellranger aggr)
\item Option 2: Importing into Seurat, then merging Seurat objects.
Barcode uniquifiction method: user-defined strings are appended to the start of the barcodes, followed by an "_", within Seurat.
}

Technical fix:
\code{importDemux2Seurat} ignores all information before a "_" in Seurat cellnames, but utilizes the "-#" information at the ends of Seurat cellnames.
\itemize{
\item Option 1: \code{importDemux2Seurat} can adjust the "-#" in the Demuxlet BARCODEs automatically for users before performing the matching step.
In order to take advantage of the automatic barcodes adjustment, just supply a vector containing the locations of the sepearate .best outputs for each lane, in the same order that lanes were combined in cellranger aggr.
\item Option 2: \code{importDemux2Seurat} ignores the user-adjustable lane information added by Seurat within cellnames.
To use with this method, run \code{importDemux2Seurat} on each lane's Seruat object separately & provide a unique name for each lane to the \code{lane.names} input, BEFORE merging into a single Seurat object.
}

Run in these ways, demuxlet information can be matched to proper cells, and lane assignments can be properly reported in the "Lane" metadata slot.
}
\examples{

pbmc <- Seurat::pbmc_small

# Load in dittoSeq's included mock, (minimal) demuxlet output.
Demux <- dittoSeq::demuxlet.example

### Method 1 ###

# Notice there is a groups metadata in this Seurat object.
getMetas(pbmc)
# We will treat these as if that holds Lane information

# Now, running importDemux2Seurat:
pbmc <- importDemux2Seurat(
    Seurat = pbmc,
    lane.meta = "groups",
    Demuxlet.best = Demux)

# Note, importDemux2Seurat can also take in the location of the .best file.
#   pbmc <- importDemux2Seurat(
#       Seurat = pbmc,
#       lane.meta = "groups",
#       Demuxlet.best = "Location/filename.best")

# demux.SNP.summary() and demux.calls.summary() can now be used.
demux.SNP.summary(pbmc)
demux.calls.summary(pbmc)

### Method 2: cellranger aggr combined data ###

# If cellranger aggr was used, lanes will be denoted by "-1", "-2", ... "-#"
#   at the ends of Seurat cellnames.
# Demuxlet should be run on each lane individually.
# Provided locations of each demuxlet.best output file, *in the same order
#   that lanes were provided to cellranger aggr* this function will then
#   adjust the "-#" within the .best BARCODEs automatically before matching
#
# pbmc <- importDemux2Seurat(
#     Seurat = pbmc,
#     Demuxlet.best = c(
#         "Location/filename1.best",
#         "Location/filename2.best"),
#     lane.names = c("g1","g2"))

}
\seealso{
Included QC visualizations:

\code{\link{demux.calls.summary}} for plotting the number of sample annotations assigned within each lane.

\code{\link{demux.SNP.summary}} for plotting the number of SNPs measured per cell.

Or, see Kang et al. Nature Biotechnology, 2018. \url{https://www.nature.com/articles/nbt.4042}. For more information about the demuxlet cell-sample deconvolution method.
}
\author{
Daniel Bunis
}
