% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDittoBulk.R
\name{importDittoBulk}
\alias{importDittoBulk}
\alias{importDittoBulk.SummarizedExperiment}
\alias{importDittoBulk.DGEList}
\title{import bulk sequencing data into a format that dittoSeq functions expect.}
\usage{
importDittoBulk(x, ...)

importDittoBulk.SummarizedExperiment(
  x,
  reductions = NULL,
  metadata = NULL,
  combine_metadata = TRUE
)

importDittoBulk.DGEList(
  x,
  reductions = NULL,
  metadata = NULL,
  combine_metadata = TRUE
)
}
\arguments{
\item{x}{placeholder}

\item{...}{For the generic, additional arguments passed to specific methods.}

\item{reductions}{placeholder}

\item{metadata}{placeholder}

\item{combine_metadata}{placeholder}
}
\value{
placeholder
One recommended assay to have, in addition to raw counts, is logcounts as this is what dittoSeq expression visualizations will look to by default.
A common way to generate one is through
}
\description{
import bulk sequencing data into a format that dittoSeq functions expect.
}
