% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{is.gene}
\alias{is.gene}
\title{Tests if an input is the name of a gene recovered in the dataset.}
\usage{
is.gene(test, object = DEFAULT)
}
\arguments{
\item{test}{"potential.gene.name" in quotes. REQUIRED.}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless DEFAULT <- object has been run.}
}
\value{
Returns TRUE if there is a row in the objects' data slot named 'test'.
}
\description{
Tests if an input is the name of a gene recovered in the dataset.
}
\examples{
pbmc <- Seurat::pbmc_small
is.gene("CD14", object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
is.gene("CD14")
}
