% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-getters.R
\name{isMeta}
\alias{isMeta}
\title{Tests if an input is the name of a meta.data slot in a target object.}
\usage{
isMeta(test, object = DEFAULT, return.values = FALSE)
}
\arguments{
\item{test}{String or vector of strings, the "potential.metadata.name"(s) to check for.}

\item{object}{A Seurat or SingleCellExperiment object to work with, OR the name of the object in "quotes".}

\item{return.values}{Logical which sets whether the function returns a logical \code{TRUE}/\code{FALSE} versus the \code{TRUE} \code{test} values . Default = \code{FALSE}
REQUIRED, unless '\code{DEFAULT <- "object"}' has been run.}
}
\value{
Returns a logical or logical vector indicating whether each instance in \code{test} is a meta.data slot within the \code{object}.
Alternatively, returns the values of \code{test} that were indeed metadata slots if \code{return.values = TRUE}.
}
\description{
Tests if an input is the name of a meta.data slot in a target object.
}
\details{
For Seurat objects, also returns TRUE for the input \code{"ident"} because, for all dittoSeq visualiztions, \code{"ident"} will retrieve a Seurat objects' clustering slot.
}
\examples{

pbmc <- Seurat::pbmc_small

# To see all metadata slots of an object
getMetas(pbmc)

# To check if something is a metadata slot
isMeta("age", object = pbmc) # False
isMeta("nCount_RNA", object = pbmc) # True

# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
isMeta("groups")

# works for multiple test metas
isMeta(c("age","nCount_RNA"))

# and with return.values = TRUE, returns all elements that are indeed slots
isMeta(c("age","nCount_RNA", "RNA_snn_res.0.8"),
    return.values = TRUE)

}
\seealso{
\code{\link{getMetas}} for returning all metadata slots of an \code{object}

\code{\link{meta}} for obtaining the contants of metadata slots
}
\author{
Daniel Bunis
}
