% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{make_hover_strings}
\alias{make_hover_strings}
\title{Outputs a string of gene expression / meta.data information for ploty hover display}
\usage{
make_hover_strings(data.hover, object, data.type = "normalized")
}
\arguments{
\item{data.hover}{The data needed in the text output. = A list of metadata names, genes, or "ident", in quotes.}

\item{object}{the Seurat or RNAseq object to draw from = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{data.type}{For when grabbing gene expression data: Should the data be "normalized" (data slot), "raw" (raw.data or counts slot), "scaled" (the scale.data slot of Seurat objects), "relative" (= pulls normalized data, then uses the scale() function to produce a relative-to-mean representation), or "normalized.to.max" (= pulls normalized data, then divides by the maximum value)? DEFAULT = "normalized"}
}
\value{
Given a list of data to grab in data.hover, outputs the 'data name': data, 'data name': data, ... for every cell of the object
}
\description{
Outputs a string of gene expression / meta.data information for ploty hover display
}
\examples{
pbmc <- Seurat::pbmc_small
make_hover_strings(c("CD34","ident","non-genes/metas-will-be-ignored"), "pbmc", "normalized")
}
