% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-getters.R
\name{meta.levels}
\alias{meta.levels}
\title{Gives the distinct values of a meta.data slot (or ident)}
\usage{
meta.levels(meta, object, cells.use = NULL)
}
\arguments{
\item{meta}{quoted "meta.data.slot" name = REQUIRED. the meta.data slot whose potential values should be retrieved.}

\item{object}{A target Seurat or SingleCellExperiment object}

\item{cells.use}{String vector of cells'/samples' names which should be included.
Alternatively, a Logical vector, the same length as the number of cells in the object, which sets which cells to include.
For the typically easier logical method, provide \code{USE} in \code{object@cell.names[USE]} OR \code{colnames(object)[USE]}).}
}
\value{
Returns the distinct values of a metadata slot given to all cells/samples or for a subset of cells/samples.
(Alternatively, returns the distinct values of clustering if \code{meta = "ident"} and the object is a \code{Seurat} object).
}
\description{
Gives the distinct values of a meta.data slot (or ident)
}
\examples{

library(Seurat)
pbmc <- pbmc_small
meta.levels("RNA_snn_res.1", object = pbmc)

}
\seealso{
\code{\link{meta}} for returning an entire metadata slots of an \code{object}, not just the potential levels

\code{\link{getMetas}} for returning all metadata slots of an \code{object}

\code{\link{isMeta}} for testing whether something is the name of a metadata slot
}
\author{
Daniel Bunis
}
