% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{meta.levels}
\alias{meta.levels}
\title{Gives the distinct values of a meta.data slot (or ident)}
\usage{
meta.levels(meta, object = DEFAULT, table.out = F)
}
\arguments{
\item{meta}{quoted "meta.data.slot" name = REQUIRED. the meta.data slot whose potential values should be retrieved.}

\item{object}{the Seurat or RNAseq object = REQUIRED, unless `DEFAULT <- "object"` has been run.}

\item{table.out}{TRUE/FALSE. Default = FALSE. Whether the numbers of incidences of each level are wanted in addition to the level names themselves.}
}
\value{
Returns the distinct values of a meta.data slot, or ident (clustering) slot if "ident" was given and the object is a Seurat object.  Can also return the counts of each as well.
}
\description{
Gives the distinct values of a meta.data slot (or ident)
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small
meta.levels("RNA_snn_res.1", object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
meta.levels("RNA_snn_res.1")
}
