% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.R
\name{meta}
\alias{meta}
\title{Returns the values of a meta.data  for all cells/samples}
\usage{
meta(meta, object = DEFAULT)
}
\arguments{
\item{meta}{quoted "meta.data" slot name = REQUIRED. the meta.data slot that should be retrieved.}

\item{object}{the Seurat or RNAseq object = REQUIRED, unless `DEFAULT <- "object"` has been run.}
}
\value{
Returns the values of a meta.data slot, or the ident (clustering) slot if "ident" was given and the object is a Seurat object.
}
\description{
Returns the values of a meta.data  for all cells/samples
}
\examples{
pbmc <- Seurat::pbmc_small
meta("res.1", object = "pbmc")
# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
meta("res.1")
}
