% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-getters.R
\name{meta}
\alias{meta}
\title{Returns the values of a meta.data  for all cells/samples}
\usage{
meta(meta, object)
}
\arguments{
\item{meta}{String, the name of the "metadata" slot to grab. OR "ident" to retireve the clustering of a Seurat \code{object}.}

\item{object}{A target Seurat or SingleCellExperiment object}
}
\value{
A named vector. Returns the values of a metadata slot, or the clustering slot if \code{meta = "ident"} and the \code{object} is a Seurat.
Names of values will be the cell/sample names.
}
\description{
Returns the values of a meta.data  for all cells/samples
}
\examples{

example(importDittoBulk, echo = FALSE)
meta("groups", object = myRNA)

}
\seealso{
\code{\link{metaLevels}} for returning just the unique discrete identities that exist within a metadata slot

\code{\link{getMetas}} for returning all metadata slots of an \code{object}

\code{\link{isMeta}} for testing whether something is the name of a metadata slot
}
\author{
Daniel Bunis
}
