% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-getters.R
\name{meta}
\alias{meta}
\title{Returns the values of a meta.data  for all cells/samples}
\usage{
meta(meta, object = DEFAULT)
}
\arguments{
\item{meta}{String, the name of the "metadata" slot to grab. OR "ident" to retireve the clustering of a Seurat \code{object}.}

\item{object}{A Seurat or SingleCellExperiment object to work with, OR the name of the object in "quotes".}
}
\value{
Returns the values of a metadata slot, or the clustering slot if \code{meta = "ident"} and the \code{object} is a Seurat.
}
\description{
Returns the values of a meta.data  for all cells/samples
}
\examples{

pbmc <- Seurat::pbmc_small
meta("RNA_snn_res.1", object = "pbmc")

# Note: if DEFAULT <- "pbmc" is run beforehand, the object input can be skipped completely.
DEFAULT <- "pbmc"
meta("RNA_snn_res.1")

}
\seealso{
\code{\link{meta.levels}} for returning just the unique discrete identities that exist within a metadata slot

\code{\link{getMetas}} for returning all metadata slots of an \code{object}

\code{\link{isMeta}} for testing whether something is the name of a metadata slot
}
\author{
Daniel Bunis
}
