% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoPlot.R
\name{multi_dittoPlot}
\alias{multi_dittoPlot}
\title{Generates multiple dittoPlots arranged into a grid.}
\usage{
multi_dittoPlot(
  object,
  vars,
  group.by,
  color.by = group.by,
  legend.show = FALSE,
  ncol = 3,
  nrow = NULL,
  main = "var",
  ylab = NULL,
  xlab = NULL,
  OUT.List = FALSE,
  ...
)
}
\arguments{
\item{object}{the Seurat or SingleCellExperiment object to draw from}

\item{vars}{c("var1","var2","var3",...). A list of vars from which to generate the separate plots}

\item{group.by}{String representing the name of a metadata to use for separating the cells/samples into discrete groups. REQUIRED.}

\item{color.by}{String representing the name of a metadata to use for setting color.}

\item{ncol}{Integer which sets how many plots will be arranged per row.  Default = 3.}

\item{nrow}{Integer which sets how many rows to arrange the plots into.  Default = NULL(/blank) --> becomes however many rows are needed to show all the data.}

\item{main, ylab}{String which sets whether / how plot titles or y-axis labels should be added to each individual plot
\itemize{
\item When set to \code{"var"}, the \code{vars} names alone will be used.
\item When set to \code{"make"}, the default dittoPlot behavior will be observed: For y-axis labels, gene vars will become "'var' expression" or similar.
\item When set as any other string, that string will be used as the title / y-axis label for every plot.
\item When set to \code{NULL}, titles / axes labels will not be added.
}}

\item{xlab, legend.show, ...}{other paramters passed along to dittoPlot.}

\item{OUT.List}{Logical. (Default = FALSE) When set to \code{TRUE}, a list of the individual plots, named by the \code{vars} being shown in each, is output instead of the combined multi-plot.}
}
\value{
Given multiple 'var' parameters, this function will output a dittoPlot for each one, arranged into a grid, with some slight tweaks to the defaults.
If \code{OUT.list} was set to TRUE, the list of individual plots, named by the \code{vars} being shown in each, is output instead of the combined multi-plot.
All parameters that can be adjusted in dittoPlot can be adjusted here.
}
\description{
Generates multiple dittoPlots arranged into a grid.
}
\examples{
library(Seurat)
pbmc <- Seurat::pbmc_small

genes <- c("CD8A","CD3E","FCER1A","CD14")
multi_dittoPlot(pbmc, genes,
    group.by = "RNA_snn_res.1", color.by = "RNA_snn_res.1")

#To make it output a grid that is 2x2, to add y-axis labels
# instead of titles, and to show legends...
multi_dittoPlot(pbmc, genes,
    group.by = "RNA_snn_res.1", color.by = "RNA_snn_res.1",
    nrow = 2, ncol = 2,           #Make grid 2x2
    main = FALSE, ylab = "make",  #Add y axis labels instead of titles
    legend.show = TRUE)           #Show legends

# To eliminate the "expression", change ylab = "var"
multi_dittoPlot(pbmc, genes,
    group.by = "RNA_snn_res.1", color.by = "RNA_snn_res.1",
    nrow = 2, ncol = 2,         #Make grid 2x2
    main = FALSE, ylab = "var", #Add y axis labels without "expression"
    legend.show = TRUE)         #Show legends

}
\seealso{
\code{\link{dittoPlot}} for the single plot version of this function
}
\author{
Daniel Bunis
}
