% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoPlot.R
\name{multi_dittoPlot}
\alias{multi_dittoPlot}
\title{Generates multiple dittoPlots arranged into a grid.}
\usage{
multi_dittoPlot(
  object,
  vars,
  group.by,
  color.by = group.by,
  legend.show = FALSE,
  ncol = 3,
  nrow = NULL,
  main = "var",
  ylab = NULL,
  xlab = NULL,
  OUT.List = FALSE,
  ...
)
}
\arguments{
\item{object}{the Seurat or SingleCellExperiment object to draw from}

\item{vars}{c("var1","var2","var3",...). A vector of gene or metadata names from which to generate the separate plots}

\item{group.by}{String representing the name of a metadata to use for separating the cells/samples into discrete groups.}

\item{color.by}{String representing the name of a metadata to use for setting color. Default = \code{group.by}.}

\item{ncol, nrow}{Integers which set how many plots will be arranged per column or per row.
Default = 3 columns aand however many rows are required.

Set both to NULL to have the grid.arrange function figure out what might be most "square" on its own.}

\item{main, ylab}{String which sets whether / how plot titles or y-axis labels should be added to each individual plot
\itemize{
\item When set to \code{"var"}, the \code{vars} names alone will be used.
\item When set to \code{"make"}, the default dittoPlot behavior will be observed: Equivalent to "make" for \code{main}, but for y-axis labels, gene vars will become "'var' expression".
\item When set as any other string, that string will be used as the title / y-axis label for every plot.
\item When set to \code{NULL}, titles / axes labels will not be added.
}}

\item{xlab, legend.show, ...}{other paramters passed along to \code{\link{dittoPlot}}.}

\item{OUT.List}{Logical. (Default = FALSE) When set to \code{TRUE}, a list of the individual plots, named by the \code{vars} being shown in each, is output instead of the combined multi-plot.}
}
\value{
Given multiple 'var' parameters, this function will output a dittoPlot for each one, arranged into a grid, just with some slight tweaks to the defaults.
If \code{OUT.list} was set to TRUE, the list of individual plots is output instead of the combined multi-plot.
All parameters that can be adjusted in dittoPlot can be adjusted here.
}
\description{
Generates multiple dittoPlots arranged into a grid.
}
\examples{
# dittoSeq handles bulk and single-cell data quit similarly.
# The SingleCellExperiment object structure is used for both,
# but all functions can be used similarly directly on Seurat
# objects as well.

example(importDittoBulk, echo = FALSE)
myRNA

genes <- getGenes(myRNA)[1:4]
multi_dittoPlot(myRNA, genes, group.by = "clustering")

# violin-plots in front is often better for large single-cell datasets,
# but we cn change the order with 'plots'
multi_dittoPlot(myRNA, genes, "clustering",
    plots = c("vlnplot","boxplot","jitter"))

#To make it output a grid that is 2x2, to add y-axis labels
# instead of titles, and to show legends...
multi_dittoPlot(myRNA, genes, "clustering",
    nrow = 2, ncol = 2,           #Make grid 2x2 (only one of these needed)
    main = NULL, ylab = "make",   #Add y axis labels instead of titles
    legend.show = TRUE)           #Show legends

# We can also facet with 'split.by'
multi_dittoPlot(myRNA, genes, "clustering",
    split.by = "SNP")

}
\seealso{
\code{\link{dittoPlot}} for the single plot version of this function
}
\author{
Daniel Bunis
}
