% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DittoSeq.References.R
\docType{data}
\name{MYcolors}
\alias{MYcolors}
\title{Protanope and Deuteranope friendly colors adapted from Wong, B. Nature Methods, 2011.}
\format{An object of class \code{character} of length 24.}
\usage{
MYcolors
}
\description{
A Character vector of 24 colors in hexadecimal form. Default color.panel for DittoSeq Visualizations

1-7 = Suggested color panel from Wong, B. Nature Methods, 2011;
8 = gray40;
9-16 = 25% darker versions of colors 1-8;
17-24 = 25% lighter versions of colors 1-8
}
\author{
Daniel Bunis
}
\keyword{datasets}
