% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkRNAseq.R
\docType{class}
\name{RNAseq-class}
\alias{RNAseq-class}
\alias{Class}
\title{The RNAseq Class}
\description{
The RNAseq object stores RNAseq data in a structure that dittoSeq plotting functions know how to handle.
All that is needed to create an RNAseq object is a DESeqDataSet, the output from the DESeq() function.
In future, outputs of Limma Voom and EdgeR will work as well.
}
\section{Slots}{

\describe{
\item{\code{counts}}{a matrix. The raw genes x samples counts data. It is recommended, but not required, that one of these should be given when a new RNBAseq object is created.}

\item{\code{DEobject}}{A DESeqDataSet. The output of having run DESeq() on your data.}

\item{\code{DEtype}}{String, the type of object. (Will become more meaningful after support for Limma and EdgeR is added)}

\item{\code{data}}{matrix, the normalized data, often the regularized log correction of the counts data generated by a call to DESeq's rlog function.}

\item{\code{meta.data}}{a data.frame containing meta-information about each sample.
Autopopulated from the DESeqDataSet upon import, or added to manually afterward.
Can be sample names, conditions, timepoints, Nreads, etc.}

\item{\code{reductions}}{a list of dimensional reductions that have been run.
Each should be a list containing 3 elements:
\itemize{
\item embeddings: a matrix containing the dimensional reduction values, with individual dimensions as the columns
\item key: How the components of this dimensionality reduction should be refered to in plot axes (ex: "PC" for pca)
\item raw.object: the original data from the initial calculation. Not used by any dittoSeq funcitons, but this is still a nice place to store the object to keep all the data together.
}}

\item{\code{samples}}{a string vector, the names of the samples.}

\item{\code{misc}}{A great place to store any other data associated with your bulk experiment that does not fit elsewhere in the object. Left \code{NULL} by default, and dittoSeq functions do not use or adjust this slot in any way.

This slot can hold data of any type, including a list of multiple data objects.}
}}

\seealso{
\code{\link{importDESeq2}} for automated import of DESeq objects into an \code{RNAseq} class object

\code{\link{addMetaRNAseq}} for how to add metadata to an \code{RNAseq} object

\code{\link{addPrcomp}} and \code{\link{addDimReduction}} for how to add dimensionality reductions to an \code{RNAseq} object for use in
\code{\link{dittoDimPlot}} visualizations.
}
